\name{comparePredictions}
\alias{comparePredictions}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Compare source predictions
}
\description{
%%  Compare source predictions from two methods, e.g. classification tree and random forest
}
\usage{
%%  comparePredictions(x)
}
\arguments{
  \item{data1}{
%%   data1:  first data set containing predictions  
}
}
\details{
%%  
}
\value{
%%  A data frame with comparisons

}
\references{
%% none
}
\author{
%%  John Karon
}
\note{
%%  none
}


\seealso{
%% 
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
