% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.2dPlot.Gauss.R
\name{fn.2dPlot.Gauss}
\alias{fn.2dPlot.Gauss}
\title{fn.2dPlot.Gauss}
\usage{
fn.2dPlot.Gauss(doc = "fn.2dPlot.Gauss", data, GroupVar, ID, Groups,
  AnalyticVars, scatterPlot = T, qqPlot = T, pvalue.digits = 3,
  Identify = F, folder = " ")
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{Name for the variable defining grouping, " " if no grouping}

\item{ID}{Name for the variable with a lab ID, " " if no lab ID is used}

\item{Groups}{Vector of values of the group variable for which plots are to be done;
if "All", use all groups; if " ", no grouping}

\item{AnalyticVars}{Vector of length two with the names of two analytic variables
to be shown in the plots}

\item{scatterPlot}{Logical (default is T): specify whether to show scatter plots when qqPlot = F}

\item{qqPlot}{Logical (default is T): specify whether to show the q-q plots with the bootstrap
envelopes and multivariate plots}

\item{pvalue.digits}{Numeric (default is 3): number of significant digits retained in tests for normality}

\item{Identify}{Logical(default is F): if T, user can identify points of interest in the plots}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
A list with the following components:
 \itemize{
\item{usage:}{  String with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed:}{ The contents of the argument data restricted to the groups used}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{analyticVars:}{  The contents of the argument AnalyticVars}
\item{params:}{  A list with the values of grouping, logical, and numberic arguments}
\item{pvalues:}{  A data frame with the p-values for univariate and bivariate tests of normality}
\item{data.check:}{ A data frame with the information on user-identified points of interest;
    value is c(NA, NA) if no data identified}
\item{location:}{ The value of the parameter folder}
}
}
\description{
Plot and check specified pairs of analysis variables for bivariate normality.
}
\section{Details}{

 If qqPlot = T, by default each page has panes in two rows and three columns
 (but there is a new page for successive groups). The function stops after producing each row
  of each plot.  Enter c ("continue") at the prompt to get the next plot.
  If qqPlot = F, only the standard qq plots are shown, and the function stops after producing each
  page.
  See the vignette for more information: visualizing each plot, the information obtained
   by using the package qqtest, the tests for bivariate normality, and identifying points of interest.
}

\examples{
data(ObsidianSources)
plot.2d.Gauss<-fn.2dPlot.Gauss(data=ObsidianSources, GroupVar="Code", ID="ID", Groups=c("A","B"),
   AnalyticVars=c("Rb","Zr"))

}
