% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.2dPlot.R
\name{fn.2dPlot}
\alias{fn.2dPlot}
\title{fn.2dPlot}
\usage{
fn.2dPlot(doc = "fn.2dPlot", data, GroupVar, ID, Groups, AnalyticVars,
  PlotByGroup = T, PlotPoints = T, LowessLine = F, Lowess.f = NA,
  KernelSmooth = F, KernelWidth = 0.3, PlotEllipses = F, PlotHulls = F,
  PlotMedians = F, Ellipses = c(0.95, 0.99), Identify = F,
  PlotColors = F, Colors = c("black", "red", "blue", "green", "purple"),
  legendLoc = "topright", folder = " ")
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{name for the variable defining grouping, " " if no grouping}

\item{ID}{name for the variable with a lab ID, " " if no lab ID is used}

\item{Groups}{vector of values of the group variable for which plots are to be done;
if "All", use all groups; grouping is required (cannot be " ").}

\item{AnalyticVars}{names of two analytic variables to be shown in the plots, vector of length 2
or matrix with 2 columns; if a matrix, the set of plots is produced for each row.}

\item{PlotByGroup}{default is T; if F, all groups are on each plot for a pair of variables}

\item{PlotPoints}{if T (the default), all points are plotted; if F, no points are plotted}

\item{LowessLine}{if T, a lowess line is plotted for each group; if F, no line is plotted}

\item{Lowess.f}{parameter for lowess()less than or equal to 1, defining the range of x-values used;
if NA (the default), uses the default value of 0.67}

\item{KernelSmooth}{if T, a kernel smooth is plotted for each group; if F (the default), no kernel smooth is plotted}

\item{KernelWidth}{the proportion of the range of x-values used in the kernel smooth; default is 0.3}

\item{PlotEllipses}{if T, Gaussian confidence ellipses are plotted for each group; if F, no ellipses are plotted}

\item{PlotHulls}{if T, the convex hull is drawn for each set of points; if F (the default), no hulls are drawn}

\item{PlotMedians}{if T, the code for each group is plotted at the median of the values for that group; default is F}

\item{Ellipses}{single value or vector of values with confidence values for the ellipses; default is c(0.95,0.99)}

\item{Identify}{if T, user can identify points of interest in the plots; default is F}

\item{PlotColors:}{if T, colors are assigned to the groups}

\item{Colors:}{single value or vector of color names; if PlotByGroup = F, the vector must have the same number of colors as the number of groups}

\item{legendLoc:}{character, location of legend for a plot with points default is "topright", alternatives are combinations of "top", "bottom", "right", "left"}

\item{folder:}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
A list with the following components:
 \itemize{
\item{usage: }{  A vector with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed: }{ The contents of the argument data restricted to the groups used}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{params.numeric: }{ A vector with the values of the arguments Lowess.f and KernelWidth}
\item{params.grouping: }{ A character vector with the values of the arguments GroupVar and Groups}
\item{ellipse.pct: }{ The value of the argument Ellipses}
\item{analyticVars: }{ The value of the argument AnalyticVars}
\item{colors:}{  A vector with the value of the argument Color}
\item{data.check: }{ If Identify = T, a data frame with the information on user-identified points
 of interest; value is c(NA,NA) if no points are identified}
\item{location:}{ The contents of the parameter folder}
 }
}
\description{
Create 2-dimensional data plot or plots.  By default, a page has panes in two rows and
two columns (but there is a new page for successive groups). The function stops after producing
each row of each plot.  Enter c ("continue") at the prompt to get the next plot.
}
\section{Details}{

 See the vignette for more information: visualizing each plot, use of colors, and
  identifying points of interest.  If Identify=T, execution stops after creating each
  bivariate plot for the user to identify any points of interest.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
#
#  plot four pairs of variables by source (default has four plots on a page)
plot.2d <- fn.2dPlot(data = ObsidianSources, GroupVar = "Code", ID = "ID", Groups = c("A","B"),
          AnalyticVars = rbind(analyticVars[1:2],analyticVars[c(1,3)], analyticVars[c(1,4)],
          analyticVars[2:3]), PlotEllipses=T, LowessLine=T)
#
#  plot one pair of variables with all sources on one plot
plot.2d <- fn.2dPlot(data = ObsidianSources, GroupVar = "Code", ID = "ID", Groups = "All",
          AnalyticVars =analyticVars[1:2], PlotByGroup=F, PlotColors=T, namesPlotEllipses=T, LowessLine=T)

}
