% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot.compare.R
\name{fn.3dPlot.compare}
\alias{fn.3dPlot.compare}
\title{fn.3dPlot.compare}
\usage{
fn.3dPlot.compare(data, GroupVar, Groups, AnalyticVars, Selections, Colors,
  SymbolSize = 0.7)
}
\arguments{
\item{data:}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar:}{name for variable defining grouping, ' ' if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done AnalyticVars: vector of names (character values) of analytic
results}

\item{Selections:}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{Colors:}{color of plotted points}

\item{SymbolSize:}{value at most 1, smaller value gives smaller diameter points}
}
\description{
create 3-dimensional data plot comparing groups
}
