% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot .R
\name{fn.3dPlot}
\alias{fn.3dPlot}
\title{fn.3dPlot}
\usage{
fn.3dPlot(data, GroupVar, Groups, AnalyticVars, Selections, ByGroup = FALSE,
  Color = "red", SymbolSize = 0.7)
}
\arguments{
\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{name for variable defining grouping, " " if no grouping}

\item{Groups}{vector of values of group variable for which plots are to be done "All" use all groups " " no grouping}

\item{AnalyticVars}{vector of names (character values) of analytic results}

\item{Selections}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{ByGroup}{if T, show scatterplot for each group for each selection of 3 variables}

\item{Color}{color of plotted points}

\item{SymbolSize}{value at most 1, smaller value gives smaller diameter points}
}
\description{
create 3-dimensional data plot
}
\examples{
data(ObsidianData)
fn.3dPlot(data = ObsidianData,
          Groups = " ",
          GroupVar = " ",
          ByGroup = FALSE,
          Selections = c(4,5,6))


}
