% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot.rotate.compare .R
\name{fn.3dPlot.rotate.compare}
\alias{fn.3dPlot.rotate.compare}
\title{fn.3dPlot.rotate.compare}
\usage{
fn.3dPlot.rotate.compare(data, GroupVar, Groups, AnalyticVars, Selections,
  Colors)
}
\arguments{
\item{GroupVar:}{name for variable defining grouping, ' ' if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Selections:}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{Color:}{color of plotted points}
}
\description{
create 3-dimensional data plot(s) comparing groups that
can be rotated data: R matrix or data frame containing the data to be analyzed
}
