% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot.rotate .R
\name{fn.3dPlot.rotate}
\alias{fn.3dPlot.rotate}
\title{fn.3dPlot.rotate}
\usage{
fn.3dPlot.rotate(data, GroupVar, Groups, AnalyticVars, Selections, ByGroup,
  Color = "red")
}
\arguments{
\item{data:}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar:}{name for variable defining grouping, " " if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
"All": use all groups
" ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Selections:}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{ByGroup:}{if T, show scatterplot for each group for each selection of 3 variables}

\item{Color:}{color of plotted points}
}
\description{
create 3-dimensional data plot(s) that can be rotated
}
