% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CheckData .R
\name{fn.CheckData}
\alias{fn.CheckData}
\title{fn.CheckData}
\usage{
fn.CheckData(doc = "fn.CheckData", data, CheckDupVars, GroupVar, Groups,
  AnalyticVars, folder, ds.duplicates, ds.NegValues, ds.Nsamples, ds.summary)
}
\arguments{
\item{data:}{R object (data frame) containing analytic data}

\item{CheckDupVars:}{vector with names of identifying variables, typically group and lab ID}

\item{GroupVar:}{if there are groups, name of variable defining the groups, default value of " ": no grouping}

\item{Groups:}{character vector of groups by which numbers of samples and summary statistics will be tabulated default value of " ": tabulations are done for the entire data set value = "All": tabulation for each distinct code in GroupVar}

\item{AnalyticVars:}{character vector of names of analytic variables for which tabulations are done}

\item{folder:}{path to folder in which to store excel files with results (e.g. "C:\\project\\results\\")}

\item{ds.duplicates:}{file name for excel file with duplicate records, with extension csv (e.g. "duplicates.csv")}

\item{ds.Nsamples:}{file name for excel file with numbers of samples, with extension csv (e.g. "Nresults.csv")}

\item{ds.summary:}{file with summary statistics for each analytic variable check matrix stored in the object data for duplicates based on variables in the argument fields}
}
\value{
If folder != " ", four excel files with duplicate observations, observations with negative
      values for one or more analytic variables, numbers of observations for each analytic
      variable, and summary statistics (quartiles and number missing)
      if Groups != " ", numbers of observations and summary statistics are by group
  A list with the following components:
     fcn.date.ver: a string with the contents of the parameter doc, date run, R version used
     dataUsed: the data frame specified by the parameter data
     params:
     analyticVars: the vector of names specified by the parameter AnalyticVars
     Duplicates: a data frame containing the observations with duplicate values
     NegativeValues: a data frame containing the observations with at least one negative value
         for a variable in AnalyticVars
     Nvalues: a data frame contain the number of observations with a value for each analytic variable
     Summary: a data frame containing the summary statistics (by group, if Group is specified)
     if folder != " ": excel.files, a list containing the folder name and the name of each of
        the excel files

  USAGE
}
\description{
basic data check: duplicate records, negative analytic values, numbers of analytic results, ranges of results
}
