% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.data .R
\name{fn.data}
\alias{fn.data}
\title{fn.data
create data frame of form to be used for all sources}
\usage{
fn.data(path, data, Group, Subset, ID, Elements)
}
\arguments{
\item{path}{to the folder
containing the data}

\item{data}{= excel file containing the data for the group}

\item{Group}{= character code for the group}

\item{Subset}{= codes for subsets of samples, ' ' if no
subsets}

\item{ID}{= character IDs for samples}

\item{Elements}{= character names for elements used, the names in the excel file}
}
\description{
fn.data
create data frame of form to be used for all sources
}
