% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.PairsPlot .R
\name{fn.PairsPlot}
\alias{fn.PairsPlot}
\title{fn.PairsPlot}
\usage{
fn.PairsPlot(data, GroupVar, Groups, AnalyticVars, Span = 2/3)
}
\arguments{
\item{data:}{R matrix or data frame containing the data to be analyzedfn}

\item{GroupVar:}{name for variable defining grouping, " " if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
"All": use all groups
" ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Span:}{value >0, <1 to define proportion of data used to estimate robust smooth}
}
\description{
pairs plots of specified analytic values, by specified groups
}
