% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca.evaluation.R
\name{fn.pca.evaluation}
\alias{fn.pca.evaluation}
\title{fn.pca.evaluation}
\usage{
fn.pca.evaluation(doc = "fn.pca.evaluation", SourceData, ArtifactData,
  labID = " ", SourceGroup, ArtifactGroup, known.sources, predicted.sources,
  AnalyticVars, Identify = F, loc.legend = "topright", folder = " ",
  ds.importance, ds.pts.outside, ds.in.out, ds.identified)
}
\arguments{
\item{doc:}{documentation in list returned by function}

\item{SourceData:}{data from known sources, including code for location and element analyses}

\item{ArtifactData:}{corresponding data from artifacts}

\item{labID:}{ID for sample, " " if none (default value)}

\item{SourceGroup:}{name of variable with code of location}

\item{ArtifactGroup:}{name of variable with code for predicted source}

\item{known.sources:}{vector of locations to be considered as sources
predicted.soures: vector of predicted sources, names must match those in known.sources if both are the same (not all need to be in known.sources)
predicted.sources: predicted sources to be compared}

\item{AnalyticVars:}{elements used in principal component analyses}

\item{loc.legend:}{location of legend added to plots (alternates are "topleft",
  "bottomright","bottomleft")
@param Identify: if T, the user can identify artifacts of interest and obtain a data set with information on those artifacts
  (default is F)}

\item{folder:}{path to folder containing result files each file name should end with .csv}

\item{ds.importance:}{name of file with percent of variance explained for the known source analysis}

\item{ds.pts.outside:}{name of file with information on artifacts with principal component pointsoutside of hull for predicted source}

\item{ds.in.out:}{table with number of artifacts, by whether inside or outside hull for predicted  source, for each predicted source}
}
\value{
The function produces two plots: the convex hulls of the first two principal components of the source data,
 and a plot with those convex hulls and the artifact data (using the weights obtained from the source data).
 The function returns a list with the following components:
\itemize{
  \item{usage:}{  A vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{  The contents of the argument data restricted to the groups used}
  \item{params}{  A list with the values of the arguments GroupVar and Groups}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{table.in.out:} {  A data frame with counts of the numbers of artifacts inside and outside of the source locations}
  \item{pts.outside:}  {  A data frame with the data for artifact points located outside of a source}
  \item{data.check:}{  If Identify=T, a data frame with the observations in dataUsed identified as of interest}
  \item{files:}{  If folder != " ", a list with path and data set names to the excel files containing importance,
  ds.in.out, pts.outside, and, if Identify=T, data for the points of interest}
   }
}
\description{
Create plots to evaluate the predicted artifact sources using the source data principal component loadings
}
\examples{
data(ObsidianSources)
data(ObsidianArtifacts)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
sources <- unique(ObsidianSources[,"Code"])
pca.eval <- fn.pca.evaluation(SourceData=ObsidianSources, ArtifactData=ObsidianArtifacts,SourceGroup= "Code",
   ArtifactGroup="Code",known.sources=sources, predicted.sources=sources, AnalyticVars=analyticVars)

}
