% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca.Gauss.R
\name{fn.pca.Gauss}
\alias{fn.pca.Gauss}
\title{fn.PCA.Gauss}
\usage{
fn.pca.Gauss(doc = "fn.pca.Gauss", data, GroupVar = "Code",
  Groups = "All", AnalyticVars, QQtest = T, folder, ds.pvalues)
}
\arguments{
\item{data:}{R matrix or data
frame containing the data to be analyzed}

\item{GroupVar:}{name for variable defining
grouping, ' ' if no grouping}

\item{Groups:}{vector of values of group variable for which
plots are to be done 'All': use all groups grouping is assumed}

\item{AnalyticVars:}{vector
of names (character values) of analytic results}

\item{folder:}{folder in which excel file
is to be stored}

\item{ds.pvalues:}{excel file with Anderson-Darling and Mardia p-values,
extension .csv create dataset data. Used based on grouping}
}
\description{
check whether first two components are Gaussian
}
