% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca.R
\name{fn.pca}
\alias{fn.pca}
\title{fn.PCA}
\usage{
fn.pca(doc = "fn.pca", data, GroupVar, Groups, AnalyticVars, ScreePlot = F,
  BoxPlots = F, PlotPoints = F, PlotEllipses = F,
  legendLoc = "topright", PlotHull = T, PlotMedians = T,
  Ellipses = c(0.95, 0.99), PlotColors = T, Colors = c("red", "black",
  "blue", "green", "purple"), Identify = F, folder = " ", ds.weights,
  ds.importance)
}
\arguments{
\item{doc:}{documentation in the list returned, default is the function name}

\item{data:}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar:}{name for variable defining grouping; if " ", no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
if "All": use all groups; if " ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Ellipses:}{value or vector of proportions for confidence ellipses
default is c(.95,.99) to produce 95% and 99% confidence ellipses}

\item{legendLoc:}{character, location of legend for a plot with points
default is "topright", alternatives are combinations of "top", "bottom",
"right", "left"}

\item{PlotColors:}{T, use list of colors in Colors for points
F, plot points as black}

\item{Colors:}{vector of color names}

\item{Identify:}{if T, the user can identify points of interest in plots; information on
these points is saved to the file; default is F}

\item{folder:}{folder in which excel files are to be stored}

\item{ds.weights:}{excel file with principal component loadings, extension.csv}

\item{ds.importance:}{excel file with percent of variation explained, extension.csv}
}
\description{
compute principal components using standardized data
}
