% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.randomForest .R
\name{fn.randomForest}
\alias{fn.randomForest}
\title{fn.randomForest}
\usage{
fn.randomForest(doc = "fn.randomForest", data, GroupVar = "Code",
  Groups = "All", AnalyticVars, Ntrees = 500, NvarUsed = 3,
  Seed = 11111, folder, ds.importance, ds.confusion)
}
\arguments{
\item{doc}{.}

\item{data}{.}

\item{GroupVar:}{name of variable defining groups, grouping is required}

\item{Groups:}{vector of
codes for groups to be used, 'All' if use all groups}

\item{AnalyticVars:}{vector with
names of analytic variables}

\item{Ntrees:}{number of trees grown, default value of 500 is
that for the randomForest function}

\item{NvarUsed:}{if not NA, number of variables to use
in each random forest call to rpart}

\item{Seed:}{if not NA, random number generator seed
to produce reproducible results}

\item{folder:}{path to folder containing excel files, must
end with '\\'}

\item{ds.importance:}{excel file with importance measures, extension .csv}

\item{ds.confusion:}{excel file with confusion matrix, extension.csv}
}
\description{
random forest analysis doc: documentation data: analysis data
}
