% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ObsidianData}
\alias{ObsidianData}
\title{Obsidian Data}
\format{An object of class \code{data.frame} with 705 rows and 15 columns.}
\usage{
ObsidianData
}
\description{
A dataset containing elemtal concentations for obsidian artefacts
}
\keyword{datasets}
