% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ObsidianSources}
\alias{ObsidianSources}
\title{Documentation for data sets}
\format{
A data frame with 113 rows and 7 variables:
\describe{
  \item{Code}{A one-character variable with source code: A, B, C, D, E, F}
  \item{ID}{A character variable with an artificial lab ID; the first character
  is the value of Code; the remaining characters are 1:n, with n samples from a source}
  \item{Rb}{The rubidium value}
  \item{Sr}{The strontium value}
  \item{Y}{The ytterbium value}
  \item{Zr}{The zirconium value}
  \item{Nb}{The niobium value}
  }
}
\source{
Data are from Steve Shackley's website http://www.swxrflab.net/ and from
  files provided by Shackley.
}
\usage{
ObsidianSources
}
\description{
Chemical composition of obsidian sources
}
\details{
A data set containing data on five elements from 113 obsidian samples from five sources in the Jemez
caldera in north central New Mexico.  The element presence values (parts per million) were obtained using
x-ray fluorescence at a laboratory at the University of California in Berkeley.
}
\keyword{datasets}
