% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_2dPlot.R
\name{ps_2dPlot}
\alias{ps_2dPlot}
\title{ps_2dPlot}
\usage{
ps_2dPlot(doc = "ps_2dPlot", data, GroupVar, ID, Groups, AnalyticVars,
  PlotByGroup = TRUE, PlotPoints = TRUE, LowessLine = FALSE,
  Lowess_f = NA, KernelSmooth = FALSE, KernelWidth = 0.3,
  PlotEllipses = FALSE, PlotHulls = FALSE, PlotMedians = FALSE,
  Ellipses = c(0.95, 0.99), Identify = FALSE, PlotColors = FALSE,
  Colors = c("black", "red", "blue", "green", "purple"),
  legendLoc = "topright", folder = " ")
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{A matrix or data frame containing the data to be analyzed}

\item{GroupVar}{The name for the variable defining grouping, " " if no grouping}

\item{ID}{The name for the variable with a lab ID, " " if no lab ID is used}

\item{Groups}{A vector of values of the group variable for which plots are to be done;
if "All", use all groups; grouping is required (cannot be " ")_}

\item{AnalyticVars}{The names of two analytic variables to be shown in the plots, vector of length 2
or matrix with 2 columns; if a matrix, the set of plots is produced for each row_}

\item{PlotByGroup}{Logical.  The default is TRUE; if FALSE, all groups are on each plot for a pair of variables}

\item{PlotPoints}{Logical.  If TRUE (the default), all points are plotted; if FALSE, no points are plotted}

\item{LowessLine}{Logical. If TRUE, a lowess line is plotted for each group; if FALSE, no line is plotted}

\item{Lowess_f}{parameter for lowess() less than or equal to 1, defining the range of x-values used;
if NA (the default), uses the default value of 0_67}

\item{KernelSmooth}{Logical.  If TRUE, a kernel smooth is plotted for each group;
if FALSE (the default), no kernel smooth is plotted}

\item{KernelWidth}{the proportion of the range of x-values used in the kernel smooth; default is 0_3}

\item{PlotEllipses}{Logical.  If TRUE, Gaussian confidence ellipses are plotted for each group;
if F (the default), no ellipses are plotted}

\item{PlotHulls}{if TRUE, the convex hull is drawn for each set of points; if FALSE (the default),
no hulls are drawn}

\item{PlotMedians}{if TRUE, the code for each group is plotted at the median of the values
for that group; default is FALSE}

\item{Ellipses}{single value or vector of values with confidence values for the ellipses; default is c(0_95,0_99)}

\item{Identify}{if TRUE, user can identify points of interest in the plots; default is FALSE}

\item{PlotColors}{Logical.  If TRUE, colors are assigned to the groups}

\item{Colors}{single value or vector of color names; if PlotByGroup = FALSE,
the vector must have the same number of colors as the number of groups}

\item{legendLoc}{Character, location of legend for a plot with points;
default is "topright", alternatives are combinations of "top", "bottom", "right", "left"}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
A list with the following components:
 \itemize{
\item{usage: }{  A string with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed: }{ The contents of the argument data restricted to the groups used}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{params: }{ A list with the values of the grouping, logical and numeric arguments}
\item{analyticVars: }{ The value of the argument AnalyticVars}
\item{colors:}{  A vector with the value of the argument Color}
\item{dataCheck: }{ If Identify = TRUE, a data frame with the information on user-identified points
 of interest; value is c(NA,NA) if no points are identified}
\item{location:}{ The path to a folder in which results will be saved}
 }
}
\description{
Create 2-dimensional data plot or plots.  By default, a page has panes in two rows and
two columns (but there is a new page for successive groups). The function stops after producing
each row of each plot.  Enter c ("continue") at the prompt to get the next plot.
}
\section{Details}{

 See the vignette for more information: visualizing each plot, use of colors, and
  identifying points of interest.  If Identify=TRUE, execution stops after creating each
  bivariate plot for the user to identify any points of interest.
}

\examples{
#
#  plot two pair of variables by source (default has four plots on a page)
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plotVars<-rbind(analyticVars[1:2], analyticVars[c(1,5)])
plot_2d <- ps_2dPlot(data = ObsidianSources, GroupVar = "Code", ID = "ID", Groups = c("A","B"),
          AnalyticVars=plotVars, PlotByGroup=TRUE, PlotColors=TRUE, PlotEllipses=TRUE,
          LowessLine=TRUE)
#
#  plot one pair of variables with all sources on one plot
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plot_2d <- ps_2dPlot(data = ObsidianSources, GroupVar = "Code", ID = "ID", Groups = "All",
          AnalyticVars =analyticVars[1:2], PlotByGroup=FALSE, PlotColors=TRUE, PlotEllipses=TRUE,
          LowessLine=TRUE)

}
