% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_2dPlotGauss.R
\name{ps_2dPlotGauss}
\alias{ps_2dPlotGauss}
\title{ps_2dPlotGauss}
\usage{
ps_2dPlotGauss(doc = "ps_2dPlotGauss", data, GroupVar, ID, Groups,
  AnalyticVars, analysisVars, scatterPlot = TRUE, QQPlot = TRUE,
  pvalue_digits = 3, Identify = FALSE, folder = " ")
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{A matrix or data frame containing the data to be analyzed}

\item{GroupVar}{The name for the variable defining grouping, " " if no grouping}

\item{ID}{The name for the variable with a lab ID, " " if no lab ID is used}

\item{Groups}{A vector of values of the group variable for which plots are to be done;
if "All", use all groups; if " ", no grouping}

\item{AnalyticVars}{A vector with the names of all analytic variables of interest}

\item{analysisVars}{A vector of length 2 with the names of the variables to be shown in the plots}

\item{scatterPlot}{Logical (default is TRUE): specify whether to show scatter plots when QQPlot = FALSE}

\item{QQPlot}{Logical (default is TRUE): specify whether to show the q-q plots with the bootstrap
envelopes and multivariate plots}

\item{pvalue_digits}{Numeric (default is 3): number of significant digits retained in tests for normality}

\item{Identify}{Logical(default is FALSE): if TRUE, user can identify points of interest in the plots}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\description{
Plot and check specified pairs of analysis variables for bivariate normality.
}
\section{Details}{

 If QQPlot = TRUE, by default each page has panes in two rows and three columns
 (but there is a new page for successive groups). The function stops after producing each row
  of each plot.  Enter c ("continue") at the prompt to get the next plot.
  If QQPlot = FALSE, only the standard qq plots are shown, and the function stops after producing each
  page.
  See the vignette for more information: visualizing each plot, the information obtained
   by using the package qqtest, the tests for bivariate normality, and identifying points of interest.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Nb","Zr","Y","Sr")
plot_2d_Gauss<-ps_2dPlotGauss(data=ObsidianSources, GroupVar="Code", ID="ID", Groups=c("A","B"),
   AnalyticVars= analyticVars, analysisVars=c("Rb","Zr"))

}
