% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_boxPlots.R
\name{ps_boxPlots}
\alias{ps_boxPlots}
\title{ps_boxPlots}
\usage{
ps_boxPlots(doc = "ps_boxPlots", data, GroupVar, Groups, AnalyticVars, Nrow,
  Ncol)
}
\arguments{
\item{doc}{A string documenting use, included in list returned; default value is the function name}

\item{data}{A data frame or matrix with data to be analyzed}

\item{GroupVar}{The name for the variable defining grouping (required)}

\item{Groups}{A vector of values of group variable for which plots are to be done; if 'All', use all groups:
if ' ', no grouping}

\item{AnalyticVars}{A vector of names (character values) of analytic results to be analyzed}

\item{Nrow}{The number of rows of plots for each page}

\item{Ncol}{The number of columns of plots for each page}
}
\value{
A list with the following components:
 \itemize{
\item{usage:}{  A string with the value of the argument doc, date run, version of R used}
\item{dataUsed:}{  A data frame with data restricted to data used for the box plots}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{params:}{  A list with elements containing the values of the grouping and numeric parameters}
\item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
}
}
\description{
Box plots of specified analytic values, by specified groups_
}
\section{DETAILS}{

If the function creates more than one plot, the code stops after each is displayed; enter c
(continue) at the prompt to display the next plot.  If Groups specifies only one group, the code
for that group is not displayed under a plot.  If an observation has a variable value that is
 missing, the variables without missing values are used in the corresponding box plots.
}

\examples{
data(ObsidianSources)
boxPlots<-ps_boxPlots(data=ObsidianSources, GroupVar="Code", Groups="All",
AnalyticVars=c("Rb","Nb", "Zr","Sr"),Nrow=2,Ncol=2)

# side-by-side box plots of each source and artifacts assigned to that source
data(ObsidianSources)
data(ObsidianArtifacts)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
ObsidianSources<-ObsidianSources[,c("Code",analyticVars)]
Artifacts <- ObsidianArtifacts[,c("Code",analyticVars)]
SourcesCode <- as.vector(ObsidianSources[,"Code"], mode="character")
ArtifactsCode <- as.vector(paste(Artifacts[,"Code"],"A",sep="_"),mode="character")
Codes <- c(SourcesCode, ArtifactsCode)
SourcesArtifacts <- data.frame(rbind(ObsidianSources,Artifacts)[,analyticVars], Code = Codes)
boxPlots<-ps_boxPlots(data=SourcesArtifacts, GroupVar="Code", Groups="All",AnalyticVars="Rb",Nrow=1,Ncol=1)

}
