% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_checkData.R
\name{ps_checkData}
\alias{ps_checkData}
\title{ps_checkData}
\usage{
ps_checkData(doc = "pc_checkData", data, CheckDupVars, GroupVar = " ",
  Groups, ID = " ", AnalyticVars, folder = " ")
}
\arguments{
\item{doc}{A character string written to the output list; default is the function name}

\item{data}{An R object (data frame) containing analytic data}

\item{CheckDupVars}{A vector with names of identifying variables, typically group and lab ID}

\item{GroupVar}{If there are groups, name of variable defining the groups,
default value of " ": no grouping}

\item{Groups}{A character vector of groups by which numbers of samples and statistics will be
tabulated.  Default is " ": tabulations are done for the entire data set.
If value = "All": tabulation for each distinct code in GroupVar}

\item{ID}{The name of lab ID, default is " " (no lab ID)}

\item{AnalyticVars}{A character vector of names of analytic variables for which tabulations are done}

\item{folder}{The path to the folder in which data frames will be saved; default is " ": no path}
}
\description{
Data checks and summaries: duplicate records, negative analytic values,
 numbers of analytic results, percentiles of results
}
\section{Detail}{

AnalyticVars must be a vector of length at least 2.  If Groups specifies selected groups (is
not equal to "All" or " "), it must be a vector of length at least 2.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
dataCheck<-ps_checkData(data=ObsidianSources,CheckDupVars=analyticVars,GroupVar="Code",Groups="All",
ID = "ID", AnalyticVars=analyticVars)

}
