% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_convexHull.R
\name{ps_convexHull}
\alias{ps_convexHull}
\title{ps_convexHull}
\usage{
ps_convexHull(data, groupVar, hullGroup)
}
\arguments{
\item{data}{R matrix or data frame containing the data to be analyzed}

\item{groupVar}{The name for variable defining grouping; a group variable is required}

\item{hullGroup}{A code defining the group for which a hull is to be plotted}
}
\value{
The function returns a matrix with the points defining the corners of the
convex hull.
}
\description{
Compute the corners of a convex hull and plot the hull: used in ps_pcaEvaluation()
}
\section{Details}{

The function assumes that the data frame data (an argument) contains group codes in
GroupVar and two principal components, named pc1 and pc2.
}

