% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_pairsPlot.R
\name{ps_pairsPlot}
\alias{ps_pairsPlot}
\title{ps_pairsPlot}
\usage{
ps_pairsPlot(doc = "ps_pairsPlot", data, GroupVar, Groups, AnalyticVars,
  Span = 2/3)
}
\arguments{
\item{doc}{A string documenting use added to the output list, default is the function name}

\item{data}{A matrix or data frame containing the data to be analyzed}

\item{GroupVar}{The name for variable defining grouping (required)}

\item{Groups}{A vector of values of group variable for which plots are to be done_
if "All": use all groups; if " ": no grouping}

\item{AnalyticVars}{A vector of names (character values) of analytic results}

\item{Span}{A value >0, <=1 defining the proportion of data used to estimate the lowess smooth_
The default value (2/3) is the default value for the lowess function_}
}
\value{
A set of pairs plots as described above and a list with the following components:
\itemize{
\item{usage:}{  A vector with the value of the argument doc, date run, version of R used}
\item{dataUsed:}{  A data frame with the observations in data restricted to the groups analyzed}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{analyticVars:}{  The vector specified by the parameter AnalyticVars}
\item{params:}{  A list with the values of the grouping and numeric arguments}
\item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
}
}
\description{
Pairs plots of specified analytic values, by specified groups
}
\section{DETAILS}{

 The function produces a pairs plot with a lowess smooth through the scatter plot for
  each pair of variables in AnalyticVars.  If Groups != " ", there is a scatter plot for
  each group in Groups.  If Groups=" ", there is one pairs plot with the data for all groups.
  Executing the function produces warnings ("span is not a graphical parameter") that can be
  ignored (changing the value of Span does change the lowess smooths).
As coded, in RStudio all plots are produced without a pause; use the back arrow in the plot pane to see
 the plots.  In base R, remove the comment symbol (#) from the browser command at the end of the final loop,
 so that the function will stop after producing each plot.
}

\examples{

data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
pairsPlot <- ps_pairsPlot(data=ObsidianSources, GroupVar="Code", Groups="All",
AnalyticVars=analyticVars)

}
