% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_pca.R
\name{ps_pca}
\alias{ps_pca}
\title{ps_pca}
\usage{
ps_pca(doc = "ps_pca", data, ID = " ", GroupVar, Groups, AnalyticVars,
  ScreePlot = FALSE, BoxPlots = FALSE, pcPlot = TRUE, PlotPoints = TRUE,
  PlotEllipses = TRUE, PlotHull = FALSE, PlotMedians = FALSE,
  Ellipses = c(0.95, 0.99), PlotColors = T, legendLoc = "topright",
  Colors = c("red", "black", "blue", "green", "purple"), Identify = FALSE,
  digits = 3, Seed = 11111, folder = " ")
}
\arguments{
\item{doc}{A string with documentation in the list returned, default
is the function name}

\item{data}{A matrix or data frame containing the data to be analyzed}

\item{ID}{An optional name for an ID, default is " " if no ID}

\item{GroupVar}{The name for variable defining grouping; a variable
name is required}

\item{Groups}{Character-valued defining the values of the group variable
for which plots are to be done.  Options are a vector of values;
 "All" (use all groups).  One of these is required}

\item{AnalyticVars}{A vector of names (character values) of analytic results}

\item{ScreePlot}{Logical, if TRUE create a scree plot, default is FALSE}

\item{BoxPlots}{Logical, if TRUE, create box plots of the first two components,
default is FALSE}

\item{pcPlot}{Logical, if TRUE (the default), create the plot of the first
two components}

\item{PlotPoints}{Logical, if TRUE (the default) and pcPlot=T,
plot the points for the first two components}

\item{PlotEllipses}{Logical, if T (the default), plot the confidence ellipse
or ellipses for each group}

\item{PlotHull}{Logical, if TRUE, plot the convex hull for each group,
default is F}

\item{PlotMedians}{Logical, if TRUE, plot the symbol for each group at the
median point for that group, default is FALSE}

\item{Ellipses}{A value or vector of proportions for confidence ellipses;
default is c(_95,_99) to produce 95\% and 99\% confidence ellipses}

\item{PlotColors}{Logical.  If TRUE, use list of colors in Colors for points;
if F, plot points as black}

\item{legendLoc}{Character, location of legend for a plot with points;
default is "topright", alternatives are combinations of "top", "bottom",
 "right", "left"}

\item{Colors}{A vector of color names; default is a vector with five names}

\item{Identify}{Logical.  If TRUE, the user can identify points of interest in plots;
information on these points is saved to a file; default is FALSE}

\item{digits}{The number of significant digits to return in objects in data frames,
default is 3}

\item{Seed}{If not NA, the seed for the random number generator used if
missing data are imputed; default is 11111}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
The function produces a plot of the first two principal components,
the contents of which are defined by the arguments PlotPoints, PlotEllipses, PlotHull,
 and PlotMedians. A scree plot and box plots are produced if requested.
 The function returns a list with the following components:
\itemize{
  \item{usage:}{  A string with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{  The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params:}{  A list with the values of the arguments for grouping, logical parameters,
  Ellipses, and Colors}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{ellipse_pct:}{  The value of the argument Ellipses}
  \item{variances:}{  A data frame including the percent of variation explained by each
  principal component and the cumulative percent explained}
  \item{weights:}{  A data frame with the principal component weights for each observation}
  \item{Predicted:}{  A data frame with the predicted values for each principal component,
   plus the value of Groups and an integer GroupIndex (with values 1:number of Groups)}
  \item{DataPlusPredicted:}{  A data frame with the data used to compute the principal components,
   plus GroupIndex (as defined above) and predicted values for each principal component}
  \item{dataCheck:}{  If Identify=TRUE, a data frame with the observations in dataUsed
   identified as of interest}
  \item{location:}{  The value of the parameter folder}
 }
}
\description{
Compute and plot principal components after standardizing the data
}
\section{Details}{

If Identify=TRUE, the user must interact with each plot (or pane,
 if there is more than one pane on a plot). To identify a point,
 place the cursor as close as possible to the point and left click;  repeat if desired.
To go to the next pane, right click and select "Stop" in base R;
click on "Finish" in the plot pane in Rstudio.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
save_pca <- ps_pca(data=ObsidianSources, ID="ID", GroupVar="Code",
Groups="All", AnalyticVars=analyticVars)

}
