% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_pcaGaussian.R
\name{ps_pcaGaussian}
\alias{ps_pcaGaussian}
\title{ps_pcaGaussian}
\usage{
ps_pcaGaussian(
  doc = "ps_pcaGaussian",
  data,
  GroupVar,
  Groups,
  gaussID = " ",
  AnalyticVars,
  qqPlot = TRUE,
  gaussIdentify = FALSE,
  folder = " "
)
}
\arguments{
\item{doc}{Documentation for the analysis, default if the function name}

\item{data}{An R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{The name for variable defining grouping; a group variable is required}

\item{Groups}{A vector of values of group variable for which plots are to be done;
if "All"', use all groups}

\item{gaussID}{An optional name for an ID, default is " " if no ID}

\item{AnalyticVars}{A vector of names (character values) of analytic results}

\item{qqPlot}{Logical, should Q-Q plots (univariate with the bootstrap envelope, multivariate)
be shown; default is TRUE}

\item{gaussIdentify}{Logical, should user identify points of interest, default is FALSE}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
The function returns a list with the following components:

\itemize{
  \item{usage:}{ A vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{ The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params_grouping:}{ A list with the values of the arguments GroupVar and Groups}
  \item{analyticVars:}{ A vector with the value of the argument AnalyticVars}
  \item{params_logical:}{ The value of QQtest}
  \item{p_values:}{ A data frame with the p-values for the Gaussian assumptions for each
   group specified}
 \item{dataCheck:}{  A data frame with data identified as generating points of interest;
 value is NA if no points are identified}
  \item{location:}{ The value of the parameter folder}
 }
}
\description{
Check whether first two principal components are Gaussian
}
\section{Details}{

This function uses the function ps_2dPlotGauss().  The function produces p-values
 from univariate and multivariate tests of normality.  It produces Q-Q plots
 of the first two principal components for each group, as well as those plots with bootstrap
 envelopes and the bivariate Q-Q plot if qqPlot=TRUE.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
pca_Gauss <- ps_pcaGaussian(data=ObsidianSources, GroupVar="Code",Groups=c("A","B"),
  AnalyticVars=analyticVars)

}
