% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test2dPlotGauss}
\alias{test2dPlotGauss}
\title{A list used to verify that ps_2dPlotGauss() is valid}
\format{A list with the following elements
\describe{
\item{usage:}{  String with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed:}{ The contents of the argument data restricted to the groups used}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{analyticVars:}{  A vector with the contents of the argument analyticVars}
\item{variablePair:}{  A vector with the contents of the argument variablePair}
\item{params:}{  A list with the values of grouping, logical, and numberic arguments}
\item{pvalues:}{  A data frame with the p-values for univariate and bivariate tests of normality}
\item{dataCheck:}{ If Identify=TRUE, a data frame with the information on user-identified points of interest}
\item{location:}{ The value of the parameter folder}
}}
\usage{
test2dPlotGauss
}
\description{
A list used to verify that ps_2dPlotGauss() is valid
}
\keyword{datasets}
