% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_pca}
\alias{test_pca}
\title{A list used to verify that ps_pca() is valid}
\format{A list with the following elements
\describe{
  \item{usage:}{  A string with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{  The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params:}{  A list with the values of the arguments for grouping, logical parameters,
  Ellipses, and Colors}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{ellipse_pct:}{  The value of the argument Ellipses}
  \item{variances:}{  A data frame including the percent of variation explained by each
  principal component and the cumulative percent explained}
  \item{weights:}{  A data frame with the principal component weights for each observation}
  \item{Predicted:}{  A data frame with the predicted values for each principal component,
   plus the value of Groups and an integer GroupIndex (with values 1:number of Groups)}
  \item{DataPlusPredicted:}{  A data frame with the data used to compute the principal components,
   plus GroupIndex (as defined above) and predicted values for each principal component}
  \item{dataCheck:}{  If Identify=TRUE, a data frame with the observations in dataUsed
   identified as of interest}
  \item{location:}{  The value of the parameter folder}
 }}
\usage{
test_pca
}
\description{
A list used to verify that ps_pca() is valid
}
\keyword{datasets}
