% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_pcaEval}
\alias{test_pcaEval}
\title{A list used to verify that ps_pcaEvaluation() is valid}
\format{A list with the following elements
\describe{
  \item{usage:}{  A vector with the contents of the argument doc, the date run,
  the version of R used}
  \item{sourceData:}{  The contents of the argument SourceData restricted to knownSources}
  \item{sourcesNA:}{ A data frame with source observations with missing data for analytic
  variables; NA if no missing data}
  \item{unknownData:}{  The contents of the argument unknownData restricted to predictedSources}
  \item{unknownsNA:}{ A data frame with unknown observations with missing data for analytic
  variables; NA if no missing data}
  \item{impError}{  Normalized root mean square error estimate for imputed data;
  NA if no missing data}
  \item{params:}{  A list with the values of the grouping and source arguments and
  values of the logical arguments}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{tableInOut:}{  A data frame with counts of the numbers of unknowns inside and
  outside of each predicted source location}
  \item{ptsOutside:}{  A data frame with the data for unknown points located outside of the
   predicted source}
  \item{dataCheck:}{If Identify=TRUE, a data frame with the observations in dataUsed identified
   as of interest; value is c(NA,NA) if no points are identified}
  \item{location:}{The value of the parameter folder}
   }}
\usage{
test_pcaEval
}
\description{
A list used to verify that ps_pcaEvaluation() is valid
}
\keyword{datasets}
