% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_pcaGauss}
\alias{test_pcaGauss}
\title{A list used to verify that ps_pcaGauss() is valid}
\format{A list with the following elements
\describe{
  \item{usage:}{ A vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{ The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params:}{ A list with the values of the arguments grouping and logical arguments}
  \item{analyticVars:}{ A vector with the value of the argument AnalyticVars}
  \item{pvalues:}{ A data frame with the p-values for the Gaussian assumptions for each
   group specified}
 \item{data_check:}{  A data frame with data identified as generating points of interest;
 value is NA if no points are identified}
  \item{location:}{ The value of the parameter folder}
 }}
\usage{
test_pcaGauss
}
\description{
A list used to verify that ps_pcaGauss() is valid
}
\keyword{datasets}
