% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_randomForest}
\alias{test_randomForest}
\title{A list used to verify that ps_randomForest() is valid}
\format{
A list with the following elements
\describe{
  \item{usage:}{ A string with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{ The contents of the argument data restricted to the groups used}
  \item{sourcesNA:}{  A data frame with data from the data frame data with missing values,
   NÁ if no missing values}
  \item{analyticVars:}{ A vector with the value of the argument AnalyticVars}
  \item{params:}{ A list with the values of the grouping, logical, and numeric arguments}
  \item{formulaRf:}{The formula used in the analysis (the variables specified in the argument AnalyticVars
   separated by + signs)}
  \item{forest:}{  A summary of the random forest call, estimated error rate, and
  confusion matrix}
  \item{importance:}{  A data frame with information on the importance of each variable
   in AnalyticVars}
  \item{confusion:}{A data frame with the estimate of the confusion matrix}
  \item{predictedData:}{  A data frame with the artifact data used for predictions; if there
  is missing data, after imputation of the missing data}
  \item{predictedNA:}{ A data frame with the observations for which missing data were imputed;
  NA if there are no missing data}
  \item{predictedSources:}{  A data frame with prediction information, sample ID (if requested),
   and values of AnalyticVars}
  \item{predictedTotals:}{  A vector with the predicted totals for each group (source)}
  \item{impError:}{ The estimated OOB (out of bag) error for imputed predictor data;
  NA if no imputed data}
  \item{location:}{ The value of the parameter folder}
 }
}
\usage{
test_randomForest
}
\description{
A list used to verify that ps_randomForest() is valid
}
\keyword{datasets}
