% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testCheckData}
\alias{testCheckData}
\title{A list used to verify that ps_checkData() is valid}
\format{A list with the following elements
\describe{
\item{usage:}{  A string with the contents of the argument doc, date run, R version used}
\item{dataUsed:}{  The data frame specified by the argument data and GroupVar}
\item{params:}{  A character vector with the values of CheckDupVars, GroupVar, and Groups}
\item{analyticVars:}{  The vector of names specified by the argument AnalyticVars}
\item{Duplicates:}{  A data frame containing the observations with duplicate values}
\item{NegativeValues:}{  A data frame containing the observations with at least one negative
 value for a variable in AnalyticVars}
\item{Nvalues:}{  A data frame contain the number of observations with a value for each analytic variable}
\item{statistics:}{  A data frame containing the statistics statistics (by group, if Group is specified)}
\item{location:}{  The value of the parameter folder}
}}
\usage{
testCheckData
}
\description{
A list used to verify that ps_checkData() is valid
}
\keyword{datasets}
