% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testCV_corr}
\alias{testCV_corr}
\title{A list used to verify that ps_CV_corr() is valid}
\format{A list with the following elements
\describe{
  \item{usage:}{  A vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{  The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params:}{  A list containing the values of the grouping,logical, and numeric parameters}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{CV:}{  A data frame with the coefficients of variation for each analytic variable in each group}
  \item{corr:}{  A data frame with the correlations between pairs of variables in each group}
  \item{location:}{  The value of the parameter folder}
      }}
\usage{
testCV_corr
}
\description{
A list used to verify that ps_CV_corr() is valid
}
\keyword{datasets}
