% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testTree}
\alias{testTree}
\title{A list used to verify that ps_tree() is valid}
\format{A list with the following elements
\describe{
  \item{usage:}{ A string with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{ The contents of the argument data restricted to the groups used}
  \item{params_grouping:}{ A list with the values of the arguments GroupVar and Groups}
  \item{analyticVars:}{ A vector with the value of the argument AnalyticVars}
  \item{params:}{ A list with the values of the grouping, logical, and splitting parameters}
  \item{model:}{ A character string with the value of the argument ModelTitle}
  \item{Tree:}{ A list with details of the tree construction_}
  \item{classification:}{A data frame showing the crossclassification of sources and predicted sources}
  \item{CpTable:}{  A data frame showing the decrease in Cp with increasing numbers of splits}
  \item{predictedSources:}{  If predictSources = T, a data frame with the predicted sources}
  \item{predictedTotals:}{  If predictedSources = T, a vector with the number of objects predicted to be from each source}
  \item{location:}{ The value of the parameter folder}
 }}
\usage{
testTree
}
\description{
A list used to verify that ps_tree() is valid
}
\keyword{datasets}
