% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.2dPlot.Gauss.R
\name{fn.2dPlot.Gauss}
\alias{fn.2dPlot.Gauss}
\title{fn.2dPlot.Gauss}
\usage{
fn.2dPlot.Gauss(doc = "fn.bivariate.Gauss", data = ObsidianSources,
  GroupVar = "Code", labID = " ", Groups = "All",
  AnalyticVars = Els[1:2], QQtest = T, pvalue.digits = 3, Identify = F,
  folder = Folder, ds.pvalues = "bivariatePvalues.csv", ds.data.check)
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{Name for the variable defining grouping, " " if no grouping}

\item{labID}{Name for the variable with a lab ID, " " if no lab ID is used}

\item{Groups}{Vector of values of the group variable for which plots are to be done;
if "All", use all groups; if " ", no grouping}

\item{AnalyticVars}{Names of two analytic variables to be shown in the plots, vector of length 2 or matrix with 2 columns;
if a matrix, the set of plots is produced for each row}

\item{QQtest}{Logical (default is T): specify whether to run the package qqtest}

\item{pvalue.digits}{Numeric (default is 3): number of significant digits retained in tests for normality}

\item{Identify}{Logical(default is F): if T, user can identify points of interest in the plots}

\item{folder}{In Windows, path to the folder containing an excel file with the identified points, ending with two forward slashes;
if " ", no data set is written}

\item{ds.identified}{Excel file name with extension .csv containing information on the identified points}
}
\value{
A list with the following components:
 \itemize{
\item{usage}{  Vector with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed}{ The contents of the argument data restricted to the groups used}
\item{analyticVars}{  The contents of the argument AnalyticVars}
\item{parameters}{  A vector with argument values for GroupVar, Groups, pvalue.digits, and QQtest}
\item{pvalues}{  A matrix with the p-values for univariate and bivariate tests of normality}
\item{data.check}{ If Identify = T, a data frame with the information on user-identified points of interest}
\item{file}{ If folder != " ", a string or list: the path to the excel file or files with pvalues and, if Identify = T,
 the information on user-identified points of interest}
}
}
\description{
Plot and check specified pairs of analysis variables for bivariate normality. The function stops after producing each plot.  Enter c ("continue") at the prompt to get the next plot.  If this function is run using Rstudio, each plot appears in a separate window, not in the Rstudio plot pane.
}
\section{Details}{

 See the vignette for more information: visualizing each plot, the information obtained by using the package qqtest,
  the tests for bivariate normality, and identifying points of interest.
}

\examples{
data(ObsidianSources)
fn.2dPlot.Gauss(data = ObsidianSources,
          Groups = " ",
          GroupVar = " ",
          ByGroup = FALSE,
          Selections = c(4,5,6))


}
