% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.2dPlot.R
\name{fn.2dPlot}
\alias{fn.2dPlot}
\title{fn.2dPlot}
\usage{
fn.2dPlot(doc = "fn.2dPlot", data, GroupVar, labID, Groups, AnalyticVars,
  PlotByGroup = T, PlotPoints = T, LowessLine = F, Lowess.f = NA,
  KernelSmooth = F, KernelWidth = 0.3, PlotEllipses = F, PlotHulls = F,
  PlotMedians = F, Ellipses = c(0.95, 0.99), Identify = F,
  PlotColors = F, Colors = c("black", "red", "blue", "green", "purple"),
  legendLoc = "topright", folder = " ", ds.identified)
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{name for the variable defining grouping, " " if no grouping}

\item{labID}{name for the variable with a lab ID, " " if no lab ID is used}

\item{Groups}{vector of values of the group variable for which plots are to be done;
if "All", use all groups; if " ", no grouping}

\item{AnalyticVars}{names of two analytic variables to be shown in the plots, vector of length 2 or matrix with 2 columns;
if a matrix, the set of plots is produced for each row}

\item{PlotByGroup}{default is T; if F, all groups are on each plot for a pair of variables}

\item{PlotPoints}{if T (the default), all points are plotted; if F, no points are plotted}

\item{LowessLine}{if T, a lowess line is plotted for each group; if F, no line is plotted}

\item{Lowess.f}{parameter for lowess()less than or equal to 1, defining the range of x-values used;
if NA (the default), uses the default value of 0.67}

\item{KernelSmooth}{if T, a kernel smooth is plotted for each group; if F (the default), no kernel smooth is plotted}

\item{KernelWidth}{the proportion of the range of x-values used in the kernel smooth; default is 0.3}

\item{PlotEllipses}{if T, Gaussian confidence ellipses are plotted for each group; if F, no ellipses are plotted}

\item{PlotHulls}{if T, the convex hull is drawn for each set of points; if F (the default), no hulls are drawn}

\item{PlotMedians}{if T, the code for each group is plotted at the median of the values for that group; default is F}

\item{Ellipses}{single value or vector of values with confidence values for the ellipses; default is c(0.95,0.99)}

\item{Identify}{if T, user can identify points of interest in the plots; default is F}

\item{folder}{in Windows, path to the folder containing an excel file with the identified points, ending with two forward slashes;
if " ", no data set is written}

\item{ds.identified}{excel file name with extension .csv containing information on the identified points}

\item{PlotColors:}{if T, colors are assigned to the groups}

\item{Colors:}{single value or vector of color names; if PlotByGroup = F, the vector must have the same number of colors as the number of groups}

\item{legendLoc:}{character, location of legend for a plot with points default is "topright", alternatives are combinations of "top", "bottom", "right", "left"}
}
\value{
A list with the following components:
 \itemize{
\item{usage: }{  A vector with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed: }{ The contents of the argument data restricted to the groups used}
\item{params.numeric: }{ A vector with the values of the arguments Lowess.f and KernelWidth}
\item{params.grouping: }{ A character vector with the values of the arguments GroupVar and Groups}
\item{ellipse.pct: }{ The value of the argument Ellipses}
\item{analyticVars: }{ The value of the argument AnalyticVars}
\item{colors:}{  A vector with the value of the argument Color}
\item{data.check: }{ If Identify = T, a data frame with the information on user-identified points of interest}
\item{folder.data.check: }{ If folder != " " and Identify = T, the path to the excel file with the information on user-identified points of interest}
}
}
\description{
Create 2-dimensional data plot or plots. The function stops after producing each plot.  Enter c ("continue") at the prompt to get the next plot.  If this function is run using Rstudio, each plot appears in a separate window, not in the Rstudio plot pane.
}
\section{Details}{

 See the vignette for more information: visualizing each plot, use of colors, and identifying points of interest.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
#
#  plot four pairs of variables by source (default has four plots on a page)
plot.2d <- fn.2dPlot(data = ObsidianSources, GroupVar = "Code", labID = "ID", Groups = c("A","B"),
          AnalyticVars = rbind(analyticVars[1:2],analyticVars[c(1,3)], analyticVars[c(1,4)],
          analyticVars[2:3]), PlotEllipses=T, LowessLine=T)
#
#  plot one pair of variables with all sources on one plot
plot.2d <- fn.2dPlot(data = ObsidianSources, GroupVar = "Code", labID = "ID", Groups = "All",
          AnalyticVars =analyticVars[1:2], PlotByGroup=F, PlotColors=T, namesPlotEllipses=T, LowessLine=T)

}
