% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot.R
\name{fn.3dPlot}
\alias{fn.3dPlot}
\title{fn.3dPlot}
\usage{
fn.3dPlot(
  doc = "fn.3dPlot",
  data,
  GroupVar,
  Groups,
  AnalyticVars,
  Selections,
  ByGroup = F,
  PlotMedians = F,
  Colors = c("red", "black", "blue", "green", "purple"),
  SymbolSize = 0.7
)
}
\arguments{
\item{doc}{A string documenting use written to the output list; default is the function name}

\item{data}{A matrix or data frame containing the data to be analyzed}

\item{GroupVar}{The name for variable defining grouping; can be " " if no grouping.}

\item{Groups}{A vector of values of group variable for which plots are to be done;
if "All", use all groups; if " ", no grouping}

\item{AnalyticVars}{A vector of names (character values) of the analytic variables}

\item{Selections}{A vector of length 3, or data frame with 3 columns, with the combinations
of the analytic variables to be plotted}

\item{ByGroup}{Logical.  If T, show scatterplot for each group for each selection of 3 variables;
if F (the default), all specified groups are on one plot}

\item{PlotMedians}{Logical.  If T, plot only the medians in each group (the points are not plotted)
if F (the default), the median locations are not plotted}

\item{Colors}{A vector with the colors of plotted points, used sequentially for the groups}

\item{SymbolSize}{A value at most 1, a smaller value gives smaller diameter points}
}
\value{
A list with the following components:
 \itemize{
\item{usage:}{  A string with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed:}{  The contents of the argument data restricted to the groups used}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{params:}{  A list with the values of the grouping, logical, numeric and Color arguments}
\item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
}
}
\description{
Create 3-dimensional data plot. The function stops after producing each plot.  Enter c ("continue") at the prompt to get the next plot.  If this function is run using Rstudio, each plot appears in a separate window, not in the Rstudio plot pane.
}
\section{Details}{

 See the vignette for more information: visualizing each plot, specification of the argument
 Selections as a matrix or data frame, and use of colors.  If the plot or plots are not by group,
  all points have the color of the first element in Colors.
  If PlotMedians = T, the value of ByGroup is not used.
}

\examples{
#  show points from several groups on one plot
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plot3d<-fn.3dPlot(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"), AnalyticVars=analyticVars,
                  Selections=rbind(analyticVars[1:3],analyticVars[2:4]))

#  plots with one group per plot
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plot3d<-fn.3dPlot(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"), ByGroup=T, AnalyticVars=analyticVars,
                  Selections=analyticVars[1:3])
}
