% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot .R
\name{fn.3dPlot}
\alias{fn.3dPlot}
\title{fn.3dPlot}
\usage{
fn.3dPlot(data, GroupVar, Groups, AnalyticVars, Selections, ByGroup = FALSE,
  Color = "red", SymbolSize = 0.7)
}
\arguments{
\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{name for variable defining grouping, " " if no grouping}

\item{Groups}{vector of values of group variable for which plots are to be done "All" use all groups " " no grouping}

\item{AnalyticVars}{vector of names (character values) of analytic results}

\item{Selections}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{ByGroup}{if T, show scatterplot for each group for each selection of 3 variables}

\item{Color}{color of plotted points}

\item{SymbolSize}{value at most 1, smaller value gives smaller diameter points


  @return   A list with the following components:
  fcn.date.ver: a vector with the contents of the argument doc, the date run, the version of R used
  dataUsed: the contents of the argument data restricted to the groups used
  params: a vector with the values of the arguments ByGroup and SymbolSize
  groups: a vector (may be of length 1) with the value of the argument Groups
  analyticVars: a vector with the value of the argument AnalyticVars
  colors: a vector with the value of the argument Colors
  dummy additional line}
}
\description{
create 3-dimensional data plot. The function stops after producing each plot.  Enter c ("continue") at the prompt to get the next plot.  If this function is run using Rstudio, each plot appears in a separate window, not in the Rstudio plot pane.
}
\examples{
data(ObsidianData)
fn.3dPlot(data = ObsidianData,
          Groups = " ",
          GroupVar = " ",
          ByGroup = FALSE,
          Selections = c(4,5,6))


}
