% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot.rotate.R
\name{fn.3dPlot.rotate}
\alias{fn.3dPlot.rotate}
\title{fn.3dPlot.rotate}
\usage{
fn.3dPlot.rotate(doc = "fn.3dPlot.rotate version 0.1", data, GroupVar, Groups,
  AnalyticVars, Selections, ByGroup = F, SymbolSize = 0.7,
  Colors = c("red", "black", "blue", "green", "purple"), folder = " ",
  ds.3dPlot, extension = "pdf")
}
\arguments{
\item{doc}{A string documenting usage written to the list return, default is the function name}

\item{data:}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar:}{name for variable defining grouping, " " if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done;
"All": use all groups;" ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Selections:}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{ByGroup:}{if T, show scatterplot for each group for each selection of 3 variables}

\item{Color:}{color of plotted points}

\item{folder:}{default is " ", when no rotated 3-d plot is saved to a file
if a rotated 3-d plot is to be saved, the path (ending in \\) to the folder in which the file is saved}

\item{ds.3dPlot:}{if folder != " ",the file name (with no extension) in which the plot is saved}

\item{extension:}{if folder != " ", the format in which the file should be saved;
options are pdf ("pdf", the default) and postscript (value is "ps")}
}
\value{
A list with the following components:
\itemize{
\item{"usage"}{a vector with the contents of the argument doc, the date run, the version of R used}
\item{"dataUsed"}{the contents of the argument data restricted to the groups used}
\item{"params"}{a vector with the values of the arguments ByGroup and SymbolSize}
\item{"groups"}{a vector (may be of length 1) with the value of the argument Groups}
\item{"analyticVars"}{a vector with the value of the argument AnalyticVars}
\item{"selections"}{a vector or matrix with the value of the argument Selections}
\item{"colors"}{a vector with the value of the argument Color}
\item{"file"}{if folder != " ": the path and file name ds.3dPlot}
}
}
\description{
Create 3-dimensional data plot(s) that can be rotated
}
\section{}{
 Details
See the vignette for details on viewing each plot, saving plots to a file, and use of colors.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plot.3d.rotate<-fn.3dPlot.rotate(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"),
                                AnalyticVars = analyticVars, Selections=analyticVars[1:3],ByGroup=T)
# two plots
plot.3d.rotate<-fn.3dPlot.rotate(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"),
                                AnalyticVars = analyticVars,
                                Selections=rbind(analyticVars[1:3],analyticVars[2:4]))
}
