% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.3dPlot.rotate.R
\name{fn.3dPlot.rotate}
\alias{fn.3dPlot.rotate}
\title{fn.3dPlot.rotate}
\usage{
fn.3dPlot.rotate(doc = "fn.3dPlot.rotate", data, GroupVar, Groups,
  AnalyticVars, Selections, ByGroup = F, ptSize = 5, Colors = c("red",
  "black", "blue", "green", "purple"), folder = " ", dsFile)
}
\arguments{
\item{doc}{A string documenting usage written to the list return, default is the function name}

\item{data:}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar:}{name for variable defining grouping, " " if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done;
"All": use all groups;" ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Selections:}{vector of length 3, or data frame with 3 columns, with combinations to be plotted}

\item{ByGroup:}{if T, show scatterplot for each group for each selection of 3 variables
default is F}

\item{ptSize:}{size of plotted points, default is 5 (a larger value gives larger points)}

\item{Color:}{color(s) of plotted points; default is a vector
red, black, blue, green, purple}

\item{folder:}{the folder to which one or more files with images will be saved;
default is " " (no files saved)}

\item{dsFile:}{the complete path to a file in folder to which each image will be saved;
if folder is not " ", this must be a valid path and file name (ends in .pdf for current function)}
}
\value{
A list with the following components:
\itemize{
\item{usage}{  A vector with the contents of the argument doc, the date run, the version of R used}
\item{dataUsed}{  The contents of the argument data restricted to the groups used}
\item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
\item{params}{  A vector with the values of the arguments ByGroup and ptRadius}
\item{groups}{  A vector (may be of length 1) with the value of the argument Groups}
\item{analyticVars}{  A vector with the value of the argument AnalyticVars}
\item{selections}{  A vector or matrix with the value of the argument Selections}
\item{colors}{  A vector with the value of the argument Color}
\item{location}{  The value of the parameter folder}
}
}
\description{
Create 3-dimensional data plot(s) that can be rotated
}
\section{}{
 Details
See the vignette for details on the use of colors.  The rotated 3d plot can be saved to a file
located at dsFile.  The code saves a file as a pdf; see the documentation for the function
rgl.postscript() for changing the format to postscript, eps, tex, or others.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plot.3d.rotate<-fn.3dPlot.rotate(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"),
                                AnalyticVars = analyticVars, Selections=analyticVars[1:3],ByGroup=T)
# two plots
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
plot.3d.rotate<-fn.3dPlot.rotate(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"),
                                AnalyticVars = analyticVars,
                                Selections=rbind(analyticVars[1:3],analyticVars[2:4]))
}
