% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.BoxPlots.R
\name{fn.BoxPlots}
\alias{fn.BoxPlots}
\title{fn.BoxPlots}
\usage{
fn.BoxPlots(doc = "fn.BoxPlots version 0.1", data, GroupVar, Groups,
  AnalyticVars, Nrow, Ncol)
}
\arguments{
\item{doc:}{string documenting use, included in list returned}

\item{data:}{data frame or matrix with data to be analyzed}

\item{GroupVar:}{name for variable defining grouping, ' ' if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done; if 'All', use all groups:
if ' ', no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results to be analyzed}

\item{Nrow:}{number of rows of plots for each page}

\item{Ncol:}{number of columns of plots for each page}
}
\value{
A list with the following components:
 \itemize{
\item{"usage"}{a vector with the value of the argument doc, date run, version of R used}
\item{"analyticVars"}{the vector specified by the parameter AnalyticVars}
\item{"params.numeric"}{a vector with the values of the arguments Nrow and Ncol}
\item{"params.grouping"}{a vector with the values of the arguments GroupVar and Groups}
\item{"analyticVars"}{a vector with the value of the argument AnalyticVars}
}

\section  DETAILS:
Only the last plot is displayed in the plot window when the function is run.  Use the back arrow
to display previous plots; click on the zoom button to create a new window with an enlarged version of the plot window.
}
\description{
Box plots of specified analytic values, by specified groups
}
\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
boxPlots<-fn.BoxPlots(data=ObsidianSources, GroupVar="Code", Groups=c("A","B"),AnalyticVars=analyticVars,Nrow=2,Ncol=2)

}
