% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.BoxPlots .R
\name{fn.BoxPlots}
\alias{fn.BoxPlots}
\title{fn.BoxPlots}
\usage{
fn.BoxPlots(data, GroupVar, Groups, AnalyticVars, Nrow, Ncol)
}
\arguments{
\item{GroupVar:}{name for variable defining grouping, ' ' if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done 'All': use all groups ' ': no
grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Nrow:}{number of rows of plots for each page}

\item{Ncol:}{number of columns of plots for each page restrict to desired set of groups}
}
\description{
box plots of specified analytic values, by specified groups
}
