% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CheckData .R
\name{fn.CheckData}
\alias{fn.CheckData}
\title{fn.CheckData}
\usage{
fn.CheckData(doc = "fn.CheckData", data, CheckDupVars, GroupVar, Groups,
  AnalyticVars, folder, ds.duplicates, ds.NegValues, ds.Nsamples, ds.summary)
}
\arguments{
\item{data:}{R object (data frame) containing analytic data}

\item{CheckDupVars:}{vector with names of identifying variables, typically group and lab ID}

\item{GroupVar:}{if there are groups, name of variable defining the groups, default value of " ": no grouping}

\item{Groups:}{character vector of groups by which numbers of samples and summary statistics will be tabulated default value of " ": tabulations are done for the entire data set value = "All": tabulation for each distinct code in GroupVar}

\item{AnalyticVars:}{character vector of names of analytic variables for which tabulations are done}

\item{folder:}{path to folder in which to store excel files with results (e.g. "C:\\project\\results\\")}

\item{ds.duplicates:}{file name for excel file with duplicate records, with extension csv (e.g. "duplicates.csv")}

\item{ds.Nsamples:}{file name for excel file with numbers of samples, with extension csv (e.g. "Nresults.csv")}

\item{ds.summary:}{file with summary statistics for each analytic variable check matrix stored in the object data for duplicates based on variables in the argument fields}
}
\description{
basic data check: duplicate records, negative analytic values, numbers of analytic results, ranges of results
}
