% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CheckData.R
\name{fn.CheckData}
\alias{fn.CheckData}
\title{fn.CheckData}
\usage{
fn.CheckData(doc = "fn.CheckData", data, CheckDupVars, GroupVar, Groups,
  AnalyticVars, folder = " ", ds.duplicates, ds.NegValues, ds.Nsamples,
  ds.summary)
}
\arguments{
\item{doc:}{a character string written to the output list; default is the function name}

\item{data:}{R object (data frame) containing analytic data}

\item{CheckDupVars:}{vector with names of identifying variables, typically group and lab ID}

\item{GroupVar:}{if there are groups, name of variable defining the groups, default value of " ": no grouping}

\item{Groups:}{character vector of groups by which numbers of samples and summary statistics will be tabulated default value of " ": tabulations are done for the entire data set value = "All": tabulation for each distinct code in GroupVar}

\item{AnalyticVars:}{character vector of names of analytic variables for which tabulations are done}

\item{folder:}{character; if " ", no files are saved; if not, path to folder in which to store excel files with results (see Detail)}

\item{ds.duplicates:}{file name for excel file with duplicate records, with extension csv (e.g. "duplicates.csv")}

\item{ds.NegValues:}{file name for excel file with observations with negative analytic values (e.g. "negValues.csv")}

\item{ds.Nsamples:}{file name for excel file with numbers of samples, with extension csv (e.g. "Nresults.csv")}

\item{ds.summary:}{file with summary statistics for each analytic variable}
}
\value{
If folder != " ", four excel files with duplicate observations, observations with negative
      values for one or more analytic variables, numbers of observations for each analytic
      variable, and summary statistics (quartiles and number missing)
      if Groups != " ", numbers of observations and summary statistics are by group
  A list with the following components:
  \itemize{
\item{"usage"}{a string with the contents of the argument doc, date run, R version used}
\item{"dataUsed"}{the data frame specified by the argument data}|
\item{"params"}{}
\item{"analyticVars"}{the vector of names specified by the argument AnalyticVars}
\item{"Duplicates"}{a data frame containing the observations with duplicate values}
\item{"NegativeValues"}{a data frame containing the observations with at least one negative value for a variable in AnalyticVars}
\item{"Nvalues"}{a data frame contain the number of observations with a value for each analytic variable}
\item{"Summary"}{a data frame containing the summary statistics (by group, if Group is specified)}
\item{"files"}{if folder != " ", a list containing the folder name and the name of each of the excel files}
        }
}
\description{
data checks and summaries: duplicate records, negative analytic values, numbers of analytic results, percentiles of results
}
\section{Detail}{

AnalyticVars must be a vector of length at least 2.
If folder = " ", no excel files are written.  If not, in windows OS, the path must end with two slashes.
}

\examples{
data("ObsidianSources")
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
CheckData<-fn.CheckData(data=ObsidianSources,CheckDupVars=c("Code","ID"),GroupVar="Code",Groups="All",AnalyticVars=analyticVars)

}
