% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CheckData.R
\name{fn.CheckData}
\alias{fn.CheckData}
\title{fn.CheckData}
\usage{
fn.CheckData(doc = "fn.CheckData", data, CheckDupVars, GroupVar = " ",
  Groups, ID = " ", AnalyticVars, folder = " ")
}
\arguments{
\item{doc:}{a character string written to the output list; default is the function name}

\item{data:}{R object (data frame) containing analytic data}

\item{CheckDupVars:}{vector with names of identifying variables, typically group and lab ID}

\item{GroupVar:}{if there are groups, name of variable defining the groups, default value of " ": no grouping}

\item{Groups:}{character vector of groups by which numbers of samples and statistics statistics will be
tabulated
 default value of " ": tabulations are done for the entire data set
 value = "All": tabulation for each distinct code in GroupVar}

\item{ID:}{name of lab ID, default is " " (no lab ID)}

\item{AnalyticVars:}{character vector of names of analytic variables for which tabulations are done}

\item{folder:}{the path to a folder in which data frames will be saved; default is " "
#``}
}
\value{
Four data frames with duplicate observations, observations with negative
      values for one or more analytic variables, numbers of observations for each analytic
      variable, and statistics statistics (quartiles and number missing)
      if Groups != " ", numbers of observations and statistics statistics are by group
  A list with the following components:
  \itemize{
\item{usage:}{  A string with the contents of the argument doc, date run, R version used}
\item{dataUsed:}{  The data frame specified by the argument data and GroupVar}|
\item{params:}{  A character vector with the values of CheckDupVars, GroupVar, and Groups}
\item{analyticVars:}{  The vector of names specified by the argument AnalyticVars}
\item{Duplicates:}{  A data frame containing the observations with duplicate values}
\item{NegativeValues:}{  A data frame containing the observations with at least one negative value for a variable in AnalyticVars}
\item{Nvalues:}{  A data frame contain the number of observations with a value for each analytic variable}
\item{statistics:}{  A data frame containing the statistics statistics (by group, if Group is specified)}
\item{location:}{  If folder != " ", the path to the folder in which data frames will be saved}
  }
}
\description{
Data checks and summaries: duplicate records, negative analytic values, numbers of analytic results, percentiles of results
}
\section{Detail}{

AnalyticVars must be a vector of length at least 2.  If Groups specifies selected groups (is
not equal to "All" or " "), if must be a vector of length at least 2.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
CheckData<-fn.CheckData(data=ObsidianSources,CheckDupVars=c("Code","ID"),GroupVar="Code",Groups="All",
  ID = "ID", AnalyticVars=analyticVars)

}
