% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CV.corr .R
\name{fn.CV.corr}
\alias{fn.CV.corr}
\title{fn.CV.corr}
\usage{
fn.CV.corr(doc = "fn.CV.corr", data, GroupVar, Groups, AnalyticVars,
  Transpose = T, folder, ds.CV, ds.corr)
}
\arguments{
\item{data:}{R matrix or data frame containing the
data to be analyzedfn}

\item{GroupVar:}{name for variable defining grouping, ' ' if no
grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
'All': use all groups ' ': no grouping AnalyticVars: vector of names (character
values) of analytic results}

\item{Transpose:}{if T, one row for the correlations between
each pair of analyses (columns are groups) if F, one row for each group, column for
correlations between a pair of analyses}

\item{folder:}{location to store excel files with
coefficients of variation and Spearman correlations}

\item{ds.CV:}{file name for
coefficients of variation, with extension .csv}

\item{ds.corr:}{file name for Spearman
correlation coefficients, with extension .csv restrict to desired set of groups}
}
\description{
Compute coefficients of variation and correlations for specified
analytic values, by specified groups
}
