% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CV.corr .R, R/fn.CV.corr.R
\name{fn.CV.corr}
\alias{fn.CV.corr}
\alias{fn.CV.corr}
\title{fn.CV.corr}
\usage{
fn.CV.corr(doc = "fn.CV.corr", data, GroupVar, Groups, AnalyticVars,
  Transpose = T, CV.digits = 2, corr.digits = 2, folder = " ", ds.CV,
  ds.corr)

fn.CV.corr(doc = "fn.CV.corr", data, GroupVar, Groups, AnalyticVars,
  Transpose = T, CV.digits = 2, corr.digits = 2, folder = " ", ds.CV,
  ds.corr)
}
\arguments{
\item{data:}{R matrix or data frame containing the data to be analyzedfn}

\item{GroupVar:}{name for variable defining grouping; if " ", if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
if "All", use all groups, if " ", no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Transpose:}{see Details}

\item{CV.digits:}{number of significant digits in CV estimates, default is 2}

\item{corr.digits:}{number of significant digits in correlation estimates, default is 2}

\item{folder:}{location to store excel files with coefficients of variation and Spearman correlations
if " " (the default), no excel files are created}

\item{ds.CV:}{file name for coefficients of variation, with extension .csv}

\item{ds.corr:}{file name for Spearman correlation coefficients, with extension .csv}

\item{data:}{R matrix or data frame containing the data to be analyzedfn}

\item{GroupVar:}{name for variable defining grouping; if " ", if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
if "All", use all groups, if " ", no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Transpose:}{see Details}

\item{CV.digits:}{number of significant digits in CV estimates, default is 2}

\item{corr.digits:}{number of significant digits in correlation estimates, default is 2}

\item{folder:}{location to store excel files with coefficients of variation and Spearman correlations
if " " (the default), no excel files are created}

\item{ds.CV:}{file name for coefficients of variation, with extension .csv}

\item{ds.corr:}{file name for Spearman correlation coefficients, with extension .csv}
}
\value{
A list with the following components:
  \item{fcn.date.ver}{a vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed}{the contents of the argument data restricted to the groups used}
  \item{params.numeric}{a vector with the values of the arguments CV.digits and corr.digits}
  \item{params.grouping}{a list with the values of the argument GroupVar and Groups}
  \item{analyticVars}{a vector with the value of the argument AnalyticVars}
  \item{CV}{a data frame with the coefficients of variation for each analytic variable in each group}
  \item{corr}{a data frame with the correlations between pairs of variables in each group}
  \item{files}{if folder != " ": a list with path and data set names to the excel files containing
      the coefficients of variations and the correlations}

A list with the following components:
  fcn.date.ver: a vector with the contents of the argument doc, the date run, the version of R used
  dataUsed: the contents of the argument data restricted to the groups used
  params.numeric: a vector with the values of the arguments CV.digits and corr.digits
  params.grouping: a list with the values of the argument GroupVar and Groups
  analyticVars: a vector with the value of the argument AnalyticVars
  CV: a data frame with the coefficients of variation for each analytic variable in each group
  corr: a data frame with the correlations between pairs of variables in each group
  if folder != " ":
    files: a list with path and data set names to the excel files containing
      the coefficients of variations and the correlations
}
\description{
Compute coefficients of variation and correlations for specified
analytic values, by specified groups

Compute coefficients of variation and correlations for specified
analytic values, by specified groups
}
\examples{


}
