% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.CV.corr.R
\name{fn.CV.corr}
\alias{fn.CV.corr}
\title{fn.CV.corr}
\usage{
fn.CV.corr(doc = "fn.CV.corr", data, GroupVar, Groups = "All", ID = " ",
  AnalyticVars, Transpose = T, CV.digits = 2, corr.digits = 2,
  plotCorrs = T, folder = " ")
}
\arguments{
\item{doc}{A string with documentation, default is the function name}

\item{data}{AR matrix or data frame containing the data to be analyzedfn}

\item{GroupVar}{The name for the variable defining grouping; if " ", no grouping}

\item{Groups}{Character valued defining the the groups used.  Options are a Vector of values
of the group variable; "All" (the default; use all groups); " ", no grouping}

\item{ID}{The name of a variable with a lab ID (used for sorting data), default is " "}

\item{AnalyticVars}{A vector of names (character values) of analytic results}

\item{Transpose}{See Details}

\item{CV.digits}{The number of significant digits in CV estimates, default is 2}

\item{corr.digits}{The number of significant digits in correlation estimates, default is 2}

\item{plotCorrs}{Logical, if T (the default), create a matrix of plots describing correlations}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
A list with the following components:
  \itemize{
  \item{usage:}{  A vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{  The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params:}{  A list containing the values of the grouping,logical, and numeric parameters}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{CV:}{  A data frame with the coefficients of variation for each analytic variable in each group}
  \item{corr:}{  A data frame with the correlations between pairs of variables in each group}
  \item{location:}{  The value of the parameter folder}
      }
}
\description{
Compute coefficients of variation and correlations for specified
analytic values, by specified groups, and plots to visualize correlations.
Observations with missing values are removed from computations using them.
}
\section{Details}{

  If Transpose=T, the correlation matrix has rows defined by the group variable and
   columns defined by the pairs of analytic variables.  If Transpose=F,
    the rows are defined by pairs of analytic variables and the columns are defined by the groups.
}

\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
CV.corr<-fn.CV.corr(data = ObsidianSources, GroupVar="Code", Groups = "All",
 AnalyticVars=analyticVars)

}
