% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.data.R
\name{fn.data}
\alias{fn.data}
\title{fn.data}
\usage{
fn.data(data, Group, Subset, ID, Elements)
}
\arguments{
\item{data}{data frame containing the data for one group}

\item{Group}{character: code for the group, added to the data set}

\item{Subset}{character: name of variable with codes for subsets of samples, " " if no subsets}

\item{ID}{character: name of IDs (typically lab IDs) for samples}

\item{Elements}{character vector: names for elements used from the data}

\item{fn.data}{create data frame with data to be analyzed, columns in specified order}
}
\value{
A data frame with columns
 \itemize{
\item{"Group"}{group code (specified in Group)}
\item{"Subset"}{subset code (specified in Subset)}
\item{"ID"}{ID value}
\item{"elements"}{one column for each element specified in Elements}
}
}
\description{
fn.data
}
\section{Details}{

This function creates a data frame with columns in a specified order.  It is useful
for combining multiple data frames into one analysis object using rbind().  The name
for a specified variable must be the same in each file used in the argument data; a
space after a variable name in an excel file imported into R yields the variable name
followed by a period.
}

\examples{
load(ObsidianData)
# create example data set by restricting to one group, removing column with group code,
# then adding that code back to the data set
fn.data(data = ObsidianData[ObsidianData[,"Code"] == "AW"][,-1]
          Group = "AW",
          Subset = " ",
          ID = "ID"
          Elements = c("Rb","Sr","Y","Zr","Nb"))

}
