% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.data.R
\name{fn.data}
\alias{fn.data}
\title{fn.data}
\usage{
fn.data(doc = "fn.data", data, Group, Subset, ID, AnalyticVars)
}
\arguments{
\item{doc}{Documentation, default is the function name}

\item{data}{Data frame containing the data for one group}

\item{Group}{Character: user-specified grouping code for these data, added to the data set}

\item{Subset}{Character: name of variable with codes for subsets of samples, " " if no subsets}

\item{ID}{Character: name of IDs (typically lab IDs) for samples}

\item{AnalyticVars}{Character vector with names for elements used from the data}
}
\value{
A list with the elements:
 \itemize{
\item{usage:}{  Documentatation, including function name, version of R used, date run)}
\item{dataUsed:}{  Data used, restricted to the specified group}
\item{analyticVars:}{  The value of AnalyticVars}
\item{dataOut:}{  The data frame created by the function, with Group, Subset (if specified), ID, and analytic variables}
}
}
\description{
Create data frame with data to be analyzed, columns in specified order
}
\section{Details}{

This function creates a data frame with columns in a specified order.  It is useful
for combining multiple data frames into one analysis object using rbind().  The name
for a specified variable must be the same in each file used in the argument data; a
space after a variable name in an excel file imported into R yields the variable name
followed by a period.
}

\examples{
# create example data set by restricting to one group, removing column with group code,
# then adding that code back to the data set
data(ObsidianSources)
fn.data(data = ObsidianSources[(ObsidianSources[,"Code"] == "A"),][,-1],
          Group = "A",
          Subset = " ",
          ID = "ID",
          AnalyticVars = c("Rb","Sr","Y","Zr","Nb"))

}
