% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.PairsPlot.R
\name{fn.PairsPlot}
\alias{fn.PairsPlot}
\title{fn.PairsPlot}
\usage{
fn.PairsPlot(doc = "fn.PairsPlot version 0.1", data, GroupVar, Groups,
  AnalyticVars, Span = 2/3)
}
\arguments{
\item{doc:}{a string documenting use added to the output list, default is the function name}

\item{data:}{R matrix or data frame containing the data to be analyzedfn}

\item{GroupVar:}{name for variable defining grouping, " " if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
"All": use all groups
" ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Span:}{value >0, <=1 defining the proportion of data used to estimate the lowess smooth}
}
\value{
A set of pairs plots as described above and a list with the following components:
      fcn.date.ver: a vector with the value of the argument doc, date run, version of R used
      data.Used: a data frame with the observations in data restricted to the groups analyzed
      analyticVars: the vector specified by the parameter AnalyticVars
      params.numeric: the value of the argument Span
      params.grouping: a vector with the values of the arguments GroupVar and Groups
      analyticVars: a vector with the value of the argument AnalyticVars
}
\description{
pairs plots of specified analytic values, by specified groups
}
\section{DETAILS}{


  The function produces a pairs plot with a lowess smooth through the scatter plot for
  each pair of variables in AnalyticVars.  If Groups != " ", there is a scatter plot for
  each group in Groups.  If Groups=" ", there is one pairs plot with the data for all groups.

  In Rstudio, each plot is in a separate window, not in the Plot pane.

  Executing the function produces warnings ("span is not a graphical parameter") that can be
  ignored (changing the value of Span does change the lowess smooths).
}

