% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca.evaluation .R
\name{fn.pca.evaluation}
\alias{fn.pca.evaluation}
\title{fn.pca.evaluation  evaluate predicted artifact sources using principal component loadings}
\usage{
fn.pca.evaluation(doc = "fn.pca.evaluation", SourceData, ArtifactData,
  SourceGroup, ArtifactGroup, known.sources, predicted.sources, Els,
  loc.legend = "topright", folder, ds.importance, ds.pts.outside, ds.in.out)
}
\arguments{
\item{doc:}{documentation in list returned by function}

\item{SourceData:}{data from known sources, including code for location and element analyses}

\item{ArtifactData:}{corresponding data from artifacts}

\item{SourceGroup:}{name of variable with code of location}

\item{ArtifactGroup:}{name of variable with code for predicted source}

\item{known.sources:}{vector of locations to be considered as sources
predicted.soures: vector of predicted sources, names must match those in known.sources if both are the same (not all need to be in known.sources)
predicted.sources: predicted sources to be compared}

\item{Els:}{elements used in principal component analyses}

\item{loc.legend:}{location of legend added to plots (alternates are "topleft",
"bottomright","bottomleft")}

\item{folder:}{path to folder containing result files each file name should end with .csv}

\item{ds.importance:}{name of file with percent of variance explained for the known source analysis}

\item{ds.pts.outside:}{name of file with information on artifacts with principal component pointsoutside of hull for predicted source}

\item{ds.in.out:}{table with number of artifacts, by whether inside or outside hull for predicted  source, for each predicted source}
}
\description{
fn.pca.evaluation  evaluate predicted artifact sources using principal component loadings
}
