% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca.evaluation.R
\name{fn.pca.evaluation}
\alias{fn.pca.evaluation}
\title{fn.pca.evaluation}
\usage{
fn.pca.evaluation(doc = "fn.pca.evaluation", SourceData, ArtifactData,
  ID = " ", SourceGroup, ArtifactGroup, known.sources, predicted.sources,
  AnalyticVars, Identify = F, loc.legend = "topright", plotAllPoints = T,
  plotHullsOutsidePoints = T, plotOutsidePoints = T, folder = " ")
}
\arguments{
\item{folder}{The path to the folder in which data frames will be saved; default is " "}

\item{doc:}{documentation in list returned by function}

\item{SourceData:}{data from known sources, including code for location and element analyses}

\item{ArtifactData:}{corresponding data from artifacts}

\item{ID:}{ID for samples (artifacts), " " if none (default value)}

\item{SourceGroup:}{name of variable with code for location}

\item{ArtifactGroup:}{name of variable with code for predicted source}

\item{known.sources:}{vector of locations to be considered as sources}

\item{predicted.soures:}{vector of predicted sources to be considered, not all need be in known.sources}

\item{AnalyticVars:}{elements used in the principal component analyses}

\item{loc.legend:}{location of legend added to plots (alternates are "topleft",
"bottomright","bottomleft")}

\item{Identify:}{if T, the user can identify artifacts of interest and obtain a data set with information on those artifacts
(default is F)}

\item{plotAllPoints:}{if T (the default), show a plot with two panes: all source points and
the convex hulls for the sources, and all unknown points with these source hulls}

\item{plotHullsOutsidePoints:}{if T (the default), show a plot with two panes: all source points and
the convex hulls for the sources, and the unknown points lying outside of their predicted source
convex hulls and these hulls}

\item{plotOutsidePoints:}{if T (the default), show a plot with one pane: athe unknown points lying
outside of their predicted source convex hulls and these hulls (the second pane for
plotHullsOutsidePoints)}
}
\value{
The function produces two plots: the convex hulls of the first two principal components
of the source data, and a plot with those convex hulls and the artifact data (using the weights
 obtained from the source data).  The function returns a list with the following components:

\itemize{
  \item{usage:}{  A vector with the contents of the argument doc, the date run,
  the version of R used}
  \item{sourceData:}{  The contents of the argument SourceData restricted to knownSources}
  \item{sourcesNA:}{ A data frame with source observations with missing data for analytic
  variables; NA if no missing data}
  \item{artifactData:}{  The contents of the argument ArtifactData restricted to predictedSources}
  \item{artifactsNA:}{ A data frame with artifact observations with missing data for analytic
  variables; NA if no missing data}
  \item{impError}{  Normalized root mean square error estimate for imputed data;
  NA if no missing data}
  \item{params:}{  A list with the values of the arguments GroupVar and Groups}
  \item{analyticVars:}{  A vector with the value of the argument AnalyticVars}
  \item{table.in.out:} {  A data frame with counts of the numbers of artifacts inside and
  outside of each predicted source location}
  \item{pts.outside:}  {  A data frame with the data for artifact points located outside of the
   predicted source}
  \item{data.check:}{  If Identify=T, a data frame with the observations in dataUsed identified
   as of interest; value is c(NA,NA) if no points are identified}
  \item{location:}{  The values of the parameter folder}
   }
}
\description{
Create plots to evaluate the predicted artifact sources using the source data principal component loadings
}
\examples{

# Evaluate predicted sources of artifacts from scatterplots

# evaluate predictions from a tree model
data(ObsidianSources)
pca.eval <- fn.pca.evaluation(SourceData=ObsidianSources,
  ArtifactData=save.tree$predictedSources, SourceGroup= "Code", ArtifactGroup="source",
  known.sources=sources, predicted.sources=sources, AnalyticVars=analyticVars, ID="ID",
  plotAllPoints=T, plotHullsOutsidePoints = T, plotOutsidePoints = T)

# evaluate predictions from a random forest analysis: plot only points outside the predicted source hull
data(ObsidianSources)
pca.eval <- fn.pca.evaluation(SourceData=ObsidianSources,
  ArtifactData=save.randomForest$predictedSources, SourceGroup= "Code", ArtifactGroup="source",
  known.sources=sources, predicted.sources=sources, AnalyticVars=analyticVars, ID="ID",
  plotAllPoints=F, plotHullsOutsidePoints = F, plotOutsidePoints = T)

}
