% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca.Gauss.R
\name{fn.pca.Gauss}
\alias{fn.pca.Gauss}
\title{fn.pca.Gauss}
\usage{
fn.pca.Gauss(doc = "fn.pca.Gauss", data, GroupVar, Groups, ID = " ",
  AnalyticVars, qqPlot = T, gaussIdentify = F, folder = " ")
}
\arguments{
\item{doc}{documentation for the analysis, default if the function name}

\item{data}{R matrix or data frame containing the data to be analyzed}

\item{GroupVar}{name for variable defining grouping; a group variable is required}

\item{Groups}{vector of values of group variable for which plots are to be done;
if "All"', use all groups}

\item{AnalyticVars}{vector of names (character values) of analytic results}

\item{qqPlot}{Logical, should Q-Q plots (univariate, multivariate) be shown, default is T}

\item{gaussIdentify}{Logical, should user identify points of interest, default is F}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}

\item{ID:}{optional name for an ID, default is " " if no ID}
}
\value{
The function produces Q-Q plots (univariate and multivariate) of the first two principal components for each group if qqPlot=T.
   The function produces p-values from univariate and multivariate tests of normality.  The function returns a list with the
   following components:
\itemize{
  \item{usage:}{ A vector with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{ The contents of the argument data restricted to the groups used}
  \item{dataNA:}{  A data frame with observations containing a least one missing value
  for an analysis variable, NA if no missing values}
  \item{params.grouping:}{ A list with the values of the arguments GroupVar and Groups}
  \item{analyticVars:}{ A vector with the value of the argument AnalyticVars}
  \item{params.logical:}{ The value of QQtest}
  \item{p.values:}{ T data frame with the p-values for the Gaussian assumptions for each group specified}
  \item{location:}{ The value of the parameter folder}
 }
}
\description{
Check whether first two principal components are Gaussian
}
\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
pca.Gauss <- fn.pca.Gauss(data=ObsidianSources, GroupVar="Code",Groups=c("A","B"), AnalyticVars=analyticVars)

}
