% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.pca .R
\name{fn.pca}
\alias{fn.pca}
\title{fn.PCA}
\usage{
fn.pca(doc = "fn.pca", data, GroupVar, Groups, AnalyticVars, ScreePlot = F,
  BoxPlots = F, PlotPoints = F, PlotEllipses = F, PlotHull = T,
  PlotMedians = T, Ellipses = c(0.95, 0.99), PlotColors = T, Colors,
  folder, ds.weights, ds.importance)
}
\arguments{
\item{data:}{R matrix or data frame containing the data to be analyzedfn}

\item{GroupVar:}{name for variable defining grouping, " " if no grouping}

\item{Groups:}{vector of values of group variable for which plots are to be done
"All": use all groups
" ": no grouping}

\item{AnalyticVars:}{vector of names (character values) of analytic results}

\item{Ellipses:}{value or vector of proportions for confidence ellipses (must be given if PlotEllipses=T)}

\item{PlotColors:}{T, use list of colors in Colors for points
F, plot points as black}

\item{Colors:}{vector of color names}

\item{folder:}{folder in which excel files are to be stored}

\item{ds.weights:}{excel file with principal component loadings, extension.csv}

\item{ds.importance:}{excel file with percent of variation explained, extension.csv

 save the results of the computations as [object name]<-fn.PCA()}
}
\description{
compute principal components using standardized data
}
