% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.randomForest.R
\name{fn.randomForest}
\alias{fn.randomForest}
\title{fn.randomForest}
\usage{
fn.randomForest(doc = "fn.randomForest", data, GroupVar, Groups = "All",
  AnalyticVars, sourceID = " ", Ntrees = 500, NvarUsed = NA,
  Seed = 11111, digitsImportance = 1, plotErrorRate = T,
  plotImportance = T, predictSources = F, predictData = NA,
  artifactID = " ", plotSourceProbs = T, folder = " ")
}
\arguments{
\item{doc}{Documentation for the function use added to model usage, default value is the function name}

\item{data}{Data frame with the data used to grow trees (source data if predictions are made)}

\item{GroupVar}{Name of variable defining groups, grouping is required}

\item{Groups}{Vector of codes for groups to be used, 'All' if use all groups}

\item{AnalyticVars}{Vector with names (character-valued) of the analytic variables}

\item{sourceID}{If not " " (the default), the name of the variable with sample ID for source data}

\item{Ntrees}{Number of trees grown, default value of 500 is that for the randomForest function}

\item{NvarUsed}{If not NA (the default), number of variables to use in each random forest call
to rpart; if NA, rpart uses the default value for randomForest()
 (the square root of the number of candidate variables)}

\item{Seed}{If not NA, a random number generator seed to produce reproducible results;
default value is 11111}

\item{digitsImportance}{Significant digits for the importance measure, default is 1}

\item{plotErrorRate}{Logical, whether to show the error rate plot, default is T}

\item{plotImportance}{Logical, whether to show the plot of variable importance, default is T}

\item{predictSources}{Logical; if T, predict sources for the data in predictData; default is F}

\item{predictData}{Data frame or matrix with data used to predict sources for observations,
must contain all variables in AnalyticVars.}

\item{artifactID}{if not " " (the default), the name of the variable with the sample ID for
artifact data}

\item{plotSourceProbs}{Logical, if T (the default) and predictSources=T, show box plots of source
probabilities}

\item{folder}{The path to the folder in which data frames will be saved; default is " "}
}
\value{
The function returns a list with the following components:

\itemize{
  \item{usage:}{ A string with the contents of the argument doc, the date run, the version of R used}
  \item{dataUsed:}{ The contents of the argument data restricted to the groups used}
  \item{sourcesNA:}{  A data frame with data from the data frame data with missing values,
   NÁ if no missing values}
  \item{analyticVars:}{ A vector with the value of the argument AnalyticVars}
  \item{params.numeric:}{ A numeric vector with the values of the arguments Ntrees, NvarUsed,\
   and Seed}
  \item{params.grouping:}{ A list with the values of the arguments GroupVar and Groups}
  \item{params.logical:}{ A vector with the values of the arguments plotErrorRate,
  plotImportance, predictSources, and plotSourceProbs}
  \item{formula.rf:}  {The formula used in the analysis (the variables specified in the argument AnalyticVars
                       separated by + signs)}
  \item{forest:}{  A summary of the random forest call, estimated error rate, and
  confusion matrix}
  \item{importance:}{  A data frame with information on the importance of each variable
   in AnalyticVars}
  \item{confusion:}{  A data frame with the estimate of the confusion matrix}
  \item{predictedData:} {  A data frame with the artifact data used for predictions; if there
  is missing data, after imputation of the missing data}
  \item{predictedNA:}{ A data frame with the observations for which missing data were imputed;
  NA if there are no missing data}
  \item{predictedSources:}{  A data frame with prediction information, sample ID (if requested),
     and values of AnalyticVars}
  \item{predictedTotals:}{  A vector with the predicted totals for each group (source)}
  \item{impError:}{ The estimated OOB (out of bag) error for imputed predictor data;
  NA if no imputed data}
  \item{location:}{ The value of the parameter folder}
 }
}
\description{
Implements a random forest analysis of source data, and predicts sources of unknowns if requested
}
\details{
The function implements a random forest analysis using the R function randomForest().
If predictSources = T and plotSourceProbs is T, the function creates two box plots.
  The first plot shows, for each source,  the set of probabilities of assignment to that source
   for the observations assigned to that source (all of these probabilities should be large).
   The second plot shows, for each source, the set of probabilities of assignment to that source
    for the observations not assigned to that source (for each source,
    there is one such probability for observation); these probabilities should be relatively small,
    and some should be zero.  See the vignette for more details and examples of these plots.
}
\examples{
data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
save.randomForest <- fn.randomForest(data=ObsidianSources, GroupVar="Code",Groups="All",
  sourceID="ID", AnalyticVars=analyticVars, NvarUsed=3, plotSourceProbs=F)
#
# predicted sources for artifacts
data(ObsidianSources)
data(ObsidianArtifacts)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
save.randomForest <- fn.randomForest(data=ObsidianSources, GroupVar="Code",Groups="All",
AnalyticVars=analyticVars, sourceID="ID", NvarUsed=3, plotErrorRate=F, plotImportance=F,
predictSources=T, predictData=ObsidianArtifacts, artifactID="ID", plotSourceProbs=T)

}
