% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.tree.R
\name{fn.tree}
\alias{fn.tree}
\title{fn.tree}
\usage{
fn.tree(doc = "fn.tree", data, GroupVar = "Code", Groups = "All",
  AnalyticVars, wts = NA, PlotTree = T, Model = "Rb+Y+Nb+Zr+Sr", folder,
  outClassify, outCpTable, outCVtable, outOptSplit)
}
\arguments{
\item{doc}{.}

\item{data}{.}

\item{GroupVar}{name of variable defining groups, grouping is required}

\item{Groups}{vector of codes for groups to be used, 'All' if use all groups}

\item{AnalyticVars}{names of analytic variables wts: option to weight the observations, if used, vector with length nrow(data) if NA, assume equal weights}

\item{wts}{.}

\item{PlotTree}{if T (true), plot the recursive partitioning tree}

\item{Model}{.}

\item{folder}{path to folder containing excel files, must end with '\\'}

\item{outClassify}{file containing results of classifying the data}

\item{outCpTable}{file containing the values of Cp at successive splits}

\item{outCVtable}{.}

\item{outOptSplit}{file containing predictors are considered in the tree model in the order in which they appeat in AnalyticVars}
}
\description{
fit recursive partitioning model doc: documentation
}
