% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ObsidianData}
\alias{ObsidianData}
\title{Documentation for data sets}
\format{An object of class \code{data.frame} with 705 rows and 15 columns.}
\usage{
ObsidianData
}
\description{
Chemical composition of obsidian sources
}
\details{
A data set
}
\keyword{datasets}
