% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_2dPlot.R, R/ps_plot.R
\name{ps_plot}
\alias{ps_plot}
\alias{ps_plot}
\title{ps_plot}
\usage{
ps_plot(data = dataUsed, groupVar = GroupVar, analyticVars = AnalyticVars,
  ps_groups = groups, ps_group, ps_group_j, groupName,
  plotEllipses = PlotEllipses, ps_ellipses = Ellipses,
  plotPoints = PlotPoints, lowessLine = LowessLine, lowess_f = Lowess_f,
  kernelSmooth = KernelSmooth, kernelWidth = Kernelwidth,
  plotHulls = PlotHulls, ps_identify = Identify)

ps_plot(data = dataUsed, groupVar = GroupVar, analyticVars = AnalyticVars,
  ps_groups = groups, ps_group, ps_group_j, groupName,
  plotEllipses = PlotEllipses, ps_ellipses = Ellipses,
  plotPoints = PlotPoints, lowessLine = LowessLine, lowess_f = Lowess_f,
  kernelSmooth = KernelSmooth, kernelWidth = Kernelwidth,
  plotHulls = PlotHulls, ps_identify = Identify)
}
\arguments{
\item{data}{A matrix or data frame containing the data to be analyzed}

\item{groupVar}{The name for the variable defining grouping}

\item{analyticVars}{The names of two analytic variables to be shown in the plots, vector of length 2
or matrix with 2 columns; if a matrix, the set of plots is produced for each row_}

\item{plotEllipses}{Logical.  If TRUE, Gaussian confidence ellipses are plotted for each group;
if F (the default), no ellipses are plotted}

\item{plotPoints}{Logical.  If TRUE (the default), all points are plotted; if FALSE, no points are plotted}

\item{lowessLine}{Logical. If TRUE, a lowess line is plotted for each group; if FALSE, no line is plotted}

\item{lowess_f}{parameter for lowess() less than or equal to 1, defining the range of x-values used;
if NA (the default), uses the default value of 0.67}

\item{kernelSmooth}{Logical.  If TRUE, a kernel smooth is plotted for each group;
if FALSE (the default), no kernel smooth is plotted}

\item{kernelWidth}{the proportion of the range of x-values used in the kernel smooth;
default is 0.3}

\item{plotHulls}{if TRUE, the convex hull is drawn for each set of points; if FALSE (the default),
no hulls are drawn}

\item{plotByGroup}{Logical.  The default is TRUE; if FALSE, all groups are on each plot for a pair of variables}

\item{ellipses}{single value or vector of values with confidence values for the ellipses; default is c(0_95,0_99)}

\item{PlotMedians}{if TRUE, the code for each group is plotted at the median of the values
for that group; default is FALSE}

\item{identify}{if TRUE, user can identify points of interest in the plots; default is FALSE}

\item{data}{A matrix or data frame containing the data to be analyzed}

\item{groupVar}{The name for the variable defining grouping}

\item{analyticVars}{The names of two analytic variables to be shown in the plots, vector of length 2
or matrix with 2 columns; if a matrix, the set of plots is produced for each row_}

\item{plotByGroup}{Logical.  The default is TRUE; if FALSE, all groups are on each plot for a pair of variables}

\item{plotPoints}{Logical.  If TRUE (the default), all points are plotted; if FALSE, no points are plotted}

\item{lowessLine}{Logical. If TRUE, a lowess line is plotted for each group; if FALSE, no line is plotted}

\item{lowess_f}{parameter for lowess() less than or equal to 1, defining the range of x-values used;
if NA (the default), uses the default value of 0.67}

\item{kernelSmooth}{Logical.  If TRUE, a kernel smooth is plotted for each group;
if FALSE (the default), no kernel smooth is plotted}

\item{kernelWidth}{the proportion of the range of x-values used in the kernel smooth;
default is 0.3}

\item{plotEllipses}{Logical.  If TRUE, Gaussian confidence ellipses are plotted for each group;
if F (the default), no ellipses are plotted}

\item{ellipses}{single value or vector of values with confidence values for the ellipses; default is c(0_95,0_99)}

\item{plotHulls}{if TRUE, the convex hull is drawn for each set of points; if FALSE (the default),
no hulls are drawn}

\item{PlotMedians}{if TRUE, the code for each group is plotted at the median of the values
for that group; default is FALSE}

\item{identify}{if TRUE, user can identify points of interest in the plots; default is FALSE}
}
\value{
If the user identifies points of interest:
 \itemize{
\item{dataCheck: }{ If ps_identify = TRUE, a data frame with the information on user-identified points
 of interest; value is c(NA,NA) if no points are identified}
 }

If the user identifies points of interest:
 \itemize{
\item{dataCheck: }{ If ps_identify = TRUE, a data frame with the information on user-identified points
 of interest; value is c(NA,NA) if no points are identified}
 }
}
\description{
Create 2-dimensional scatter plot for one group: used by ps_2dPlot().

Create 2-dimensional scatter plot for one group: used by ps_2dPlot().
}
\section{Details}{

 This function is used internally by the functions ps_2dPlot and


 This function is used internally by the functions ps_2dPlot and
}

\examples{




}
