% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plot.R
\name{ps_plot}
\alias{ps_plot}
\title{ps_plot}
\usage{
ps_plot(data, ps_groupVar, ps_byGroup, ps_plotAllGroups, useVars,
  plotPoints, lowessLine, lowess_f, plotMedians, kernelSmooth, kernelWidth,
  plotHulls, plotEllipses, ps_ellipses, groupIndex, ps_colors, ps_legend,
  ps_identify)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed}

\item{ps_groupVar}{Character.  Optional name of the grouping variable.}

\item{ps_byGroup}{Logical.  If TRUE, plots are being created by group.}

\item{ps_plotAllGroups}{Logical.  If TRUE, all groups are one one plot, with identifying information}

\item{useVars}{A vector of length 2: the names of two analytic variables to
be shown in the plots}

\item{plotPoints}{Logical.  If TRUE, all points are plotted; if FALSE, no points are plotted}

\item{lowessLine}{Logical. If TRUE, a lowess line is plotted for each group; if FALSE, no line is plotted}

\item{lowess_f}{A parameter for lowess() less than or equal to 1, defining the range of x-values used}

\item{plotMedians}{Logical. If TRUE, the median value for each group is plotted with the
code for the group}

\item{kernelSmooth}{Logical.  If TRUE, a kernel smooth is plotted for each group;
if FALSE, no kernel smooth is plotted}

\item{kernelWidth}{the proportion of the range of x-values used in the kernel smooth}

\item{plotHulls}{Logical. If TRUE, the convex hull is drawn for each set of points; if FALSE,
no hulls are drawn}

\item{plotEllipses}{Logical.  If TRUE, Gaussian confidence ellipses are plotted for each group;
if FALSE, no ellipses are plotted}

\item{ps_ellipses}{single value or vector of values with confidence values for the ellipses}

\item{groupIndex}{Integer.  Used to specify the plotting character and color for multiple
groups shown on one plot.}

\item{ps_colors}{Character vector of plotting colors for multiple groups shown on one plot}

\item{ps_legend}{Character, specifying location of legend with multiple groups shown on one plot}

\item{ps_identify}{Logical. If TRUE, user can identify points of interest in the plots}
}
\value{
If the user identifies points of interest:
 \itemize{
\item{dataCheck: }{ If ps_identify = TRUE, a data frame with the information on user-identified points
 of interest; value is c(NA,NA) if no points are identified}
 }
}
\description{
Create 2-dimensional scatter plot for one group: used by ps_2dPlot().
}
\section{Details}{

 This function is used internally by the function ps_2dPlot.  The default values are those
 from ps_2dPlot.  data will contain a group code, a lab ID (if used in the data set), and all
 analytic variables, for display if the user identifies points of interest.  If plotMedians is TRUE,
 the convex hull and median of the points are plotted for each group.
}

\examples{

data(ObsidianSources)
analyticVars<-c("Rb","Sr","Y","Zr","Nb")
ps_plot(data=ObsidianSources,useVars=c("Rb","Nb"),plotPoints=TRUE,lowessLine=TRUE,
lowess_f=NA, plotMedians=FALSE, kernelSmooth=FALSE,kernelWidth=0.3,plotHulls=FALSE,
plotEllipses=TRUE,ps_ellipses=0.95,ps_identify=FALSE)

}
