% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources.data.check}
\alias{sources.data.check}
\title{Data for sources identified as potentially of interest in a plot of
zirconium versus rubidium for Jemez source data.}
\format{A data frame with 5 rows and 7 variables:
\describe{
  \item{group}{A character with the code for the source.}
  \item{ID}{The laboratory ID for the artifact.}
  \item{Rb}{A vector with the rubidium values.}
  \item{Sr}{A vector with the strontium values.}
  \item{Y}{A vector with the yttrium values.}
  \item{Zr}{A vector with the zirconium values.}
  \item{Nb}{A vector with with the niobium values.}
  }}
\source{
The data frame was obtained from the function ps_2dPlotGauss()
using Identify=TRUE.
}
\usage{
sources.data.check
}
\description{
Data for sources identified as potentially of interest in a plot of
zirconium versus rubidium for Jemez source data.
}
\keyword{datasets}
