% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree.data.check}
\alias{tree.data.check}
\title{Data for artifacts with a predicted Jemez source likely misclassified
by a classifiication tree model.}
\format{A data frame with 11 rows and 9 variables:
\describe{
  \item{group}{A character with the code for the predicted source.}
  \item{ID}{The laboratory ID for the artifact.}
  \item{Rb}{A vector with the rubidium values.}
  \item{Sr}{A vector with with the strontium values.}
  \item{Y}{A vector with the ytterbium values.}
  \item{Zr}{A vector with the zirconium values.}
  \item{Nb}{A vector with with the niobium values.}
  \item{pc1}{A vector with the values of the first principal component
  of the principal component plot of the Jemez sources and the artifacts.}
  \item{pc1}{A vector with the values of the second principal component
  of the principal component plot of the Jemez sources and the artifacts.}
  }}
\source{
The data frame was obtained from the function ps_pcaEvaluation()
using Identify=TRUE.
}
\usage{
tree.data.check
}
\description{
A data set containing the predicted source code, lab ID, element values,
and first two principal components for these artifacts.
}
\keyword{datasets}
